Lorentz.Base

Description

Foundation of Lorentz development.

Documentation

newtype (inp :: [Type]) :-> (out :: [Type]) infixr 1

Alias for instruction which hides inner types representation via T.

Constructors

I

Fields

unI :: Instr (ToTs inp) (ToTs out)

type (%>) = (:->) infixr 1

Alias for :->, seems to make signatures more readable sometimes.

Let's someday decide which one of these two should remain.

type (&) (a :: Type) (b :: [Type]) = a ': b infixr 2

(#) :: (a :-> b) -> (b :-> c) -> a :-> c

compileLorentz :: (inp :-> out) -> Instr (ToTs inp) (ToTs out)

For use outside of Lorentz.

compileLorentzContract :: forall cp st. (NoOperation cp, NoOperation st, NoBigMap cp, CanHaveBigMap st) => Contract cp st -> Instr '[ToT (cp, st)] '[ToT ([Operation], st)]

Version of compileLorentz specialized to instruction corresponding to contract code.

interpretLorentzInstr :: (IsoValuesStack inp, IsoValuesStack out) => ContractEnv -> (inp :-> out) -> Rec Identity inp -> Either MichelsonFailed (Rec Identity out)

Interpret a Lorentz instruction, for test purposes.

parseLorentzValue :: forall v. (IsoValue v, SingI (ToT v), Typeable (ToT v)) => Text -> Either ParseLorentzError v

Parse textual representation of a Michelson value and turn it into corresponding Haskell value.

Note: it won't work in some complex cases, e. g. if there is a lambda which uses an instruction which depends on current contract's type. Obviously it can not work, because we don't have any information about a contract to which this value belongs (there is no such contract at all).

analyzeLorentz :: (inp :-> out) -> AnalyzerRes

Lorentz version of analyzer.

transformStringsLorentz :: Bool -> (MText -> MText) -> (inp :-> out) -> inp :-> out

Lorentz version of transformStrings.

type ContractOut st = '[([Operation], st)]

type Contract cp st = '[(cp, st)] :-> ContractOut st

data SomeContract where

Constructors

SomeContract :: (KnownValue cp, KnownValue st, NoOperation cp, NoOperation st, NoBigMap cp, CanHaveBigMap st) => Contract cp st -> SomeContract

type Lambda i o = '[i] :-> '[o]