Lorentz.CompileExts

Documentation

newtype LorentzCompilationWay cp st

Defines a way to turn lorentz contract into typed contract.

You probably want to choose some smart implementation for this, which may add some annotations to resulting contract.

Constructors

LorentzCompilationWay

Fields

compileLorentzContract' :: Contract cp st -> FullContract (ToT cp) (ToT st)

lcwDumb :: (Each [Typeable, SingI] [ToT cp, ToT st], Each '[NoOperation] [cp, st], NoBigMap cp, CanHaveBigMap st) => LorentzCompilationWay cp st

Attaches no additional information.

lcwEntryPoints :: LcwEntryPointsC False cp st => LorentzCompilationWay cp st

Fits for case when your contract exposes multiple entrypoints via having sum type as its parameter.

In particular, this will attach field annotations to immediate parameter "arms" which will be named as corresponding constructor names.

lcwEntryPointsRecursive :: LcwEntryPointsC True cp st => LorentzCompilationWay cp st

Similar to lcwEntryPoints, but for case of parameter being defined as several nested datatypes.

In particular, this will traverse sum types recursively, stopping at Michelson primitives (like Natural) and constructors with number of fields different from one.