Lorentz.Doc

Documentation

doc :: DocItem di => di -> s :-> s

Put a document item.

docGroup :: DocGrouping -> (inp :-> out) -> inp :-> out

Group documentation built in the given piece of code into block dedicated to one thing, e.g. to one entry point.

buildLorentzDoc :: (inp :-> out) -> ContractDoc

contractName :: Text -> (inp :-> out) -> inp :-> out

Give a name to given contract. Apply it to the whole contract code.

Re-exports

class (Typeable d, KnownNat (DocItemPosition d)) => DocItem d where

A piece of documentation describing one property of a thing, be it a name or description of a contract, or an error throwable by given endpoint.

Items of the same type appear close to each other in a rendered documentation and form a section.

Doc items are later injected into a contract code via a dedicated nop-like instruction. Normally doc items which belong to one section appear in resulting doc in the same order in which they appeared in the contract.

Minimal complete definition

docItemSectionName, docItemToMarkdown

Associated Types

type DocItemPosition d = (pos :: Nat) | pos -> d

Position of this item in the resulting documentation; the smaller the value, the higher the section with this element will be placed.

Documentation structure is not necessarily flat. If some doc item consolidates a whole documentation block within it, this block will have its own placement of items independent from outer parts of the doc.

type DocItemPlacement d :: DocItemPlacementKind

Defines where given doc item should be put. There are two options: 1. Inline right here (default behaviour); 2. Put into definitions section.

Methods

docItemSectionName :: Maybe Text

When multiple items of the same type belong to one section, how this section will be called.

If not provided, section will contain just untitled content.

docItemSectionDescription :: Maybe Builder

Description of a section.

Can be used to mention some common things about all elements of this section. Markdown syntax is permitted here.

docItemRef :: d -> DocItemRef (DocItemPlacement d)

Defines a function which constructs an unique identifier of given doc item, if it has been decided to put the doc item into definitions section.

Identifier should be unique both among doc items of the same type and items of other types. Thus, consider using "typeId-contentId" pattern.

docItemRef :: DocItemPlacement d ~ DocItemInlined => d -> DocItemRef (DocItemPlacement d)

Defines a function which constructs an unique identifier of given doc item, if it has been decided to put the doc item into definitions section.

Identifier should be unique both among doc items of the same type and items of other types. Thus, consider using "typeId-contentId" pattern.

docItemToMarkdown :: HeaderLevel -> d -> Builder

Render given doc item to Markdown, preferably one line, optionally with header.

Accepts the smallest allowed level of header. (Using smaller value than provided one will interfere with existing headers thus delivering mess).

docItemDependencies :: d -> [SomeDocDefinitionItem]

All doc items which this doc item refers to.

They will automatically be put to definitions as soon as given doc item is detected.

docItemsOrder :: [d] -> [d]

This function accepts doc items put under the same section in the order in which they appeared in the contract and returns their new desired order. It's also fine to use this function for filtering or merging doc items.

Default implementation * leaves inlined items as is; * for items put to definitions, lexicographically sorts them by their id.

class TypeHasDoc a where

Description for a Haskell type appearing in documentation.

Minimal complete definition

typeDocMdDescription

Methods

typeDocName :: Proxy a -> Text

Name of type as it appears in definitions section.

Default definition derives name from Generics. If it does not fit, consider defining this function manually. (We tried using Data for this, but it produces names including module names which is not do we want).

typeDocName :: (Generic a, KnownSymbol (GenericTypeName a)) => Proxy a -> Text

Name of type as it appears in definitions section.

Default definition derives name from Generics. If it does not fit, consider defining this function manually. (We tried using Data for this, but it produces names including module names which is not do we want).

typeDocMdDescription :: Builder

Explanation of a type. Markdown formatting is allowed.

typeDocMdReference :: Proxy a -> WithinParens -> Builder

How reference to this type is rendered, in Markdown.

Examples: * Integer, * Maybe ().

Consider using one of the following functions as default implementation; which one to use depends on number of type arguments in your type: * homomorphicTypeDocMdReference * poly1TypeDocMdReference * poly2TypeDocMdReference

If none of them fits your purposes precisely, consider using customTypeDocMdReference.

typeDocMdReference :: (Typeable a, IsHomomorphic a) => Proxy a -> WithinParens -> Builder

How reference to this type is rendered, in Markdown.

Examples: * Integer, * Maybe ().

Consider using one of the following functions as default implementation; which one to use depends on number of type arguments in your type: * homomorphicTypeDocMdReference * poly1TypeDocMdReference * poly2TypeDocMdReference

If none of them fits your purposes precisely, consider using customTypeDocMdReference.

typeDocDependencies :: Proxy a -> [SomeTypeWithDoc]

All types which this type directly contains.

Used in automatic types discovery.

typeDocDependencies :: (Generic a, GTypeHasDoc (Rep a)) => Proxy a -> [SomeTypeWithDoc]

All types which this type directly contains.

Used in automatic types discovery.

typeDocHaskellRep :: TypeDocHaskellRep a

For complex types - their immediate Haskell representation.

For primitive types set this to Nothing.

For homomorphic types use homomorphicTypeDocHaskellRep implementation.

For polymorhpic types consider using concreteTypeDocHaskellRep as implementation.

Modifier haskellRepNoFields can be used to hide names of fields, beneficial for newtypes.

Another modifier called haskellRepStripFieldPrefix can be used for datatypes to leave only meaningful part of name in every field.

typeDocHaskellRep :: (Generic a, GTypeHasDoc (Rep a), IsHomomorphic a) => TypeDocHaskellRep a

For complex types - their immediate Haskell representation.

For primitive types set this to Nothing.

For homomorphic types use homomorphicTypeDocHaskellRep implementation.

For polymorhpic types consider using concreteTypeDocHaskellRep as implementation.

Modifier haskellRepNoFields can be used to hide names of fields, beneficial for newtypes.

Another modifier called haskellRepStripFieldPrefix can be used for datatypes to leave only meaningful part of name in every field.

typeDocMichelsonRep :: TypeDocMichelsonRep a

Final michelson representation of a type.

For homomorphic types use homomorphicTypeDocMichelsonRep implementation.

For polymorhpic types consider using concreteTypeDocMichelsonRep as implementation.

typeDocMichelsonRep :: (SingI (ToT a), IsHomomorphic a) => TypeDocMichelsonRep a

Final michelson representation of a type.

For homomorphic types use homomorphicTypeDocMichelsonRep implementation.

For polymorhpic types consider using concreteTypeDocMichelsonRep as implementation.

type PolyTypeHasDocC ts = Each '[TypeHasDoc] ts

Constraint, required when deriving TypeHasDoc for polymorphic type with the least possible number of methods defined manually.

data SomeTypeWithDoc where

Data hides some type implementing TypeHasDoc.

Constructors

SomeTypeWithDoc :: TypeHasDoc td => Proxy td -> SomeTypeWithDoc

data DDescription

Description of something.

Constructors

DDescription Text

data DError

An error thrown by a contract.

Constructors

DError

data DType where

Doc element with description of a type.

Constructors

DType :: TypeHasDoc a => Proxy a -> DType

contractDocToMarkdown :: ContractDoc -> LText

Render given contract documentation to markdown document.

class HaveCommonTypeCtor a b

Require two types to be built from the same type constructor.

E.g. HaveCommonTypeCtor (Maybe Integer) (Maybe Natural) is defined, while HaveCmmonTypeCtor (Maybe Integer) [Integer] is not.

class IsHomomorphic a

Require this type to be homomorphic.

genericTypeDocDependencies :: forall a. (Generic a, GTypeHasDoc (Rep a)) => Proxy a -> [SomeTypeWithDoc]

Implement typeDocDependencies via getting all immediate fields of a datatype.

Note: this will not include phantom types, I'm not sure yet how this scenario should be handled (@martoon).

customTypeDocMdReference :: (Text, DType) -> [DType] -> WithinParens -> Builder

Render a reference to a type which consists of type constructor (you have to provide name of this type constructor and documentation for the whole type) and zero or more type arguments.

homomorphicTypeDocMdReference :: forall (t :: Type). (Typeable t, TypeHasDoc t, IsHomomorphic t) => Proxy t -> WithinParens -> Builder

Derive typeDocMdReference, for homomorphic types only.

poly1TypeDocMdReference :: forall t (r :: Type) (a :: Type). (r ~ t a, Typeable t, Each '[TypeHasDoc] [r, a], IsHomomorphic t) => Proxy r -> WithinParens -> Builder

Derive typeDocMdReference, for polymorphic type with one type argument, like Maybe Integer.

poly2TypeDocMdReference :: forall t (r :: Type) (a :: Type) (b :: Type). (r ~ t a b, Typeable t, Each '[TypeHasDoc] [r, a, b], IsHomomorphic t) => Proxy r -> WithinParens -> Builder

Derive typeDocMdReference, for polymorphic type with two type arguments, like Lambda Integer Natural.

homomorphicTypeDocHaskellRep :: forall a. (Generic a, GTypeHasDoc (Rep a)) => TypeDocHaskellRep a

Implement typeDocHaskellRep for a homomorphic type.

Note that it does not require your type to be of IsHomomorphic instance, which can be useful for some polymorhpic types which, for documentation purposes, we want to consider homomorphic. Example: Operation is in fact polymorhpic, but we don't want this fact to be reflected in the documentation.

concreteTypeDocHaskellRep :: forall a b. (Typeable a, IsoValue a, Generic a, GTypeHasDoc (Rep a), HaveCommonTypeCtor b a) => TypeDocHaskellRep b

Implement typeDocHaskellRep on example of given concrete type.

This is a best effort attempt to implement typeDocHaskellRep for polymorhpic types, as soon as there is no simple way to preserve type variables when automatically deriving Haskell representation of a type.

haskellRepNoFields :: TypeDocHaskellRep a -> TypeDocHaskellRep a

Erase fields from Haskell datatype representation.

Use this when rendering fields names is undesired.

haskellRepStripFieldPrefix :: HasCallStack => TypeDocHaskellRep a -> TypeDocHaskellRep a

Cut fields prefixes which we use according to the style guide.

E.g. cmMyField field will be transformed to myField.

homomorphicTypeDocMichelsonRep :: forall a. SingI (ToT a) => TypeDocMichelsonRep a

Implement typeDocMichelsonRep for homomorphic type.

concreteTypeDocMichelsonRep :: forall a b. (Typeable a, SingI (ToT a), HaveCommonTypeCtor b a) => TypeDocMichelsonRep b

Implement typeDocMichelsonRep on example of given concrete type.

This function exists for the same reason as concreteTypeDocHaskellRep.

Orphan instances

Each ((Typeable :: Type -> Constraint) ': (TypeHasDoc ': ([] :: [Type -> Constraint]))) (i ': (o ': ([] :: [Type]))) => TypeHasDoc (Lambda i o)

Instance details

Methods

typeDocName :: Proxy (Lambda i o) -> Text

typeDocMdDescription :: Builder

typeDocMdReference :: Proxy (Lambda i o) -> WithinParens -> Builder

typeDocDependencies :: Proxy (Lambda i o) -> [SomeTypeWithDoc]

typeDocHaskellRep :: TypeDocHaskellRep (Lambda i o)

typeDocMichelsonRep :: TypeDocMichelsonRep (Lambda i o)