Lorentz.Errors

Description

Advanced errors.

Haskell to Value conversion

class IsError e where

Haskell type representing error.

Methods

errorToVal :: e -> (forall t. KnownErrorT t => Value t -> r) -> r

Converts a Haskell error into Value representation.

errorFromVal :: (Typeable t, SingI t) => Value t -> Either Text e

Converts a Value into Haskell error.

customErrorToVal :: (LooseSumC e, HasCallStack) => e -> (forall t. KnownErrorT t => Value t -> r) -> r

Implementation of errorToVal for custom errors.

customErrorFromVal :: forall t e. (SingI t, LooseSumC e) => Value t -> Either Text e

Implementation of errorFromVal for custom errors.

type family CustomErrorNoIsoValue a where ...

Equations

CustomErrorNoIsoValue a = TypeError ((Text "No IsoValue instance for " :<>: ShowType a) :$$: Text "It has custom error representation")

deriveCustomError :: Name -> Q [Dec]

Derive IsError instance for given type.

This will also forbid deriving IsoValue instance for that type to avoid having multiple different Michelson representations.

data UnspecifiedError

Use this type as replacement for () when you really want to leave error cause unspecified.

Constructors

UnspecifiedError

Instructions

failUsing :: IsError e => e -> s :-> t

Fail with the given Haskell value.

failUsingArg :: forall err name fieldTy s s'. FailUsingArg err name fieldTy s s'

Fail with given error, picking argument for error from the top of the stack.

If your error constructor does not carry an argument, use failUsing function instead. Consider the following practice: once error datatype for your contract is defined, create a specialization of this function to the error type.

type FailUsingArg e name fieldTy s s' = (KnownSymbol name, IsError e, CtorHasOnlyField name e fieldTy, Each [Typeable, SingI] '[ToT fieldTy], HasCallStack) => Label name -> (fieldTy ': s) :-> s'

Signature of userFailWith.

failUnexpected :: HasCallStack => MText -> s :-> t

Fail, providing a reference to the place in the code where this function is called.

Like error in Haskell code, this instruction is for internal errors only.