Lorentz.Macro

Description

Common Michelson macros defined using Lorentz syntax.

Compare

type IfCmpXConstraints a op = (Typeable a, ArithOpHs Compare a a, UnaryArithOpHs op (ArithResHs Compare a a), UnaryArithResHs op (ArithResHs Compare a a) ~ Bool)

eq :: (ArithOpHs Compare n n, UnaryArithOpHs Eq' (ArithResHs Compare n n)) => (n & (n & s)) :-> (UnaryArithResHs Eq' (ArithResHs Compare n n) & s)

neq :: (ArithOpHs Compare n n, UnaryArithOpHs Neq (ArithResHs Compare n n)) => (n & (n & s)) :-> (UnaryArithResHs Neq (ArithResHs Compare n n) & s)

lt :: (ArithOpHs Compare n n, UnaryArithOpHs Lt (ArithResHs Compare n n)) => (n & (n & s)) :-> (UnaryArithResHs Lt (ArithResHs Compare n n) & s)

gt :: (ArithOpHs Compare n n, UnaryArithOpHs Gt (ArithResHs Compare n n)) => (n & (n & s)) :-> (UnaryArithResHs Gt (ArithResHs Compare n n) & s)

le :: (ArithOpHs Compare n n, UnaryArithOpHs Le (ArithResHs Compare n n)) => (n & (n & s)) :-> (UnaryArithResHs Le (ArithResHs Compare n n) & s)

ge :: (ArithOpHs Compare n n, UnaryArithOpHs Ge (ArithResHs Compare n n)) => (n & (n & s)) :-> (UnaryArithResHs Ge (ArithResHs Compare n n) & s)

ifEq0 :: IfCmp0Constraints a Eq' => (s :-> s') -> (s :-> s') -> (a & s) :-> s'

ifGe0 :: IfCmp0Constraints a Ge => (s :-> s') -> (s :-> s') -> (a & s) :-> s'

ifGt0 :: IfCmp0Constraints a Gt => (s :-> s') -> (s :-> s') -> (a & s) :-> s'

ifLe0 :: IfCmp0Constraints a Le => (s :-> s') -> (s :-> s') -> (a & s) :-> s'

ifLt0 :: IfCmp0Constraints a Lt => (s :-> s') -> (s :-> s') -> (a & s) :-> s'

ifNeq0 :: IfCmp0Constraints a Neq => (s :-> s') -> (s :-> s') -> (a & s) :-> s'

ifEq :: IfCmpXConstraints a Eq' => (s :-> s') -> (s :-> s') -> (a & (a & s)) :-> s'

ifGe :: IfCmpXConstraints a Ge => (s :-> s') -> (s :-> s') -> (a & (a & s)) :-> s'

ifGt :: IfCmpXConstraints a Gt => (s :-> s') -> (s :-> s') -> (a & (a & s)) :-> s'

ifLe :: IfCmpXConstraints a Le => (s :-> s') -> (s :-> s') -> (a & (a & s)) :-> s'

ifLt :: IfCmpXConstraints a Lt => (s :-> s') -> (s :-> s') -> (a & (a & s)) :-> s'

ifNeq :: IfCmpXConstraints a Neq => (s :-> s') -> (s :-> s') -> (a & (a & s)) :-> s'

Fail

fail_ :: a :-> c

Warning: fail_ remains in code

Analog of the FAIL macro in Michelson. Its usage is discouraged because it doesn't carry any information about failure.

Assertion macros

They differ from the same macros in Michelson, because those macros use FAIL macro which is not informative (fails with unit). If you really want Michelson versions (maybe to produce exact copy of an existing contract), you can pass UnspecifiedError, then FAILWITH will be called with unit.

assert :: IsError err => err -> (Bool & s) :-> s

assertEq0 :: (IfCmp0Constraints a Eq', IsError err) => err -> (a & s) :-> s

assertNeq0 :: (IfCmp0Constraints a Neq, IsError err) => err -> (a & s) :-> s

assertLt0 :: (IfCmp0Constraints a Lt, IsError err) => err -> (a & s) :-> s

assertGt0 :: (IfCmp0Constraints a Gt, IsError err) => err -> (a & s) :-> s

assertLe0 :: (IfCmp0Constraints a Le, IsError err) => err -> (a & s) :-> s

assertGe0 :: (IfCmp0Constraints a Ge, IsError err) => err -> (a & s) :-> s

assertEq :: (IfCmpXConstraints a Eq', IsError err) => err -> (a & (a & s)) :-> s

assertNeq :: (IfCmpXConstraints a Neq, IsError err) => err -> (a & (a & s)) :-> s

assertLt :: (IfCmpXConstraints a Lt, IsError err) => err -> (a & (a & s)) :-> s

assertGt :: (IfCmpXConstraints a Gt, IsError err) => err -> (a & (a & s)) :-> s

assertLe :: (IfCmpXConstraints a Le, IsError err) => err -> (a & (a & s)) :-> s

assertGe :: (IfCmpXConstraints a Ge, IsError err) => err -> (a & (a & s)) :-> s

assertNone :: IsError err => err -> (Maybe a & s) :-> s

assertSome :: IsError err => err -> (Maybe a & s) :-> (a & s)

assertLeft :: IsError err => err -> (Either a b & s) :-> (a & s)

assertRight :: IsError err => err -> (Either a b & s) :-> (b & s)

assertUsing :: IsError a => a -> (Bool & s) :-> s

Syntactic Conveniences

dipX :: forall (n :: Nat) inp out s s'. (DipX (ToPeano n) (Above (ToPeano n) inp) s s', (Above (ToPeano n) inp ++ s) ~ inp, (Above (ToPeano n) inp ++ s') ~ out) => (s :-> s') -> inp :-> out

DII+P macro. For example, `dipX @3` is DIIIP.

dropX :: forall (n :: Nat) a inp out s s'. (DipX (ToPeano n) (Above (ToPeano n) inp) s s', (Above (ToPeano n) inp ++ s) ~ inp, (Above (ToPeano n) inp ++ s') ~ out, s ~ (a ': s')) => inp :-> out

Custom Lorentz macro that drops element with given index (starting from 0) from the stack.

cloneX :: forall (n :: Nat) a s. CloneX (ToPeano n) a s => (a & s) :-> CloneXT (ToPeano n) a s

Duplicate the top of the stack n times.

For example, `cloneX @3` has type `a & s :-> a & a & a & a & s`.

duupX :: forall (n :: Nat) inp. DuupX (ToPeano n) inp (At (ToPeano (n - 1)) inp) => inp :-> (At (ToPeano (n - 1)) inp & inp)

DUU+P macro. For example, `duupX @3` is DUUUP, it puts the 3-rd (starting from 1) element to the top of the stack.

elevateX :: forall (n :: Nat) inp out s a. (DuupX (ToPeano (1 + n)) inp a, DipX (ToPeano (n + 1)) (Above (ToPeano (n + 1)) (a ': inp)) (a ': s) s, (a ': (Above (ToPeano n) inp ++ s)) ~ out, a ~ At (ToPeano n) inp, (Above (ToPeano (n + 1)) (a ': inp) ++ (a ': s)) ~ (a ': inp), (Above (ToPeano (n + 1)) (a ': inp) ++ s) ~ (a ': (Above (ToPeano n) inp ++ s)), ((1 + n) - 1) ~ n) => inp :-> out

Move item with given index (starting from 0) to the top of the stack.

TODO: probably it can be implemented more efficiently, so if we ever want to optimize gas consumption we can rewrite it. It only makes sense if it's applied to a relatively large index.

caar :: (((a, b1), b2) & s) :-> (a & s)

cadr :: (((a, b1), b2) & s) :-> (b1 & s)

cdar :: ((a1, (a2, b)) & s) :-> (a2 & s)

cddr :: ((a1, (a2, b)) & s) :-> (b & s)

ifRight :: ((b & s) :-> s') -> ((a & s) :-> s') -> (Either a b & s) :-> s'

ifSome :: ((a & s) :-> s') -> (s :-> s') -> (Maybe a & s) :-> s'

mapCar :: ((a & s) :-> (a1 & s)) -> ((a, b) & s) :-> ((a1, b) & s)

mapCdr :: ((b & ((a, b) & s)) :-> (b1 & ((a, b) & s))) -> ((a, b) & s) :-> ((a, b1) & s)

papair :: (a & (b & (c & s))) :-> (((a, b), c) & s)

ppaiir :: (a & (b & (c & s))) :-> ((a, (b, c)) & s)

unpair :: ((a, b) & s) :-> (a & (b & s))

setCar :: ((a, b1) & (b2 & s)) :-> ((b2, b1) & s)

setCdr :: ((a, b1) & (b2 & s)) :-> ((a, b2) & s)

setInsert :: IsComparable e => (e & (Set e & s)) :-> (Set e & s)

Insert given element into set.

This is a separate function from updateMap because stacks they operate with differ in length.

mapInsert :: (MapInstrs map, IsComparable k) => (k ': (v ': (map k v ': s))) :-> (map k v ': s)

Insert given element into map.

setInsertNew :: (IsComparable e, KnownValue err) => (forall s0. (e ': s0) :-> (err ': s0)) -> (e & (Set e & s)) :-> (Set e & s)

Insert given element into set, ensuring that it does not overwrite any existing entry.

As first argument accepts container name.

mapInsertNew :: (MapInstrs map, IsComparable k, KnownValue e) => (forall s0. (k ': s0) :-> (e ': s0)) -> (k ': (v ': (map k v ': s))) :-> (map k v ': s)

Insert given element into map, ensuring that it does not overwrite any existing entry.

As first argument accepts container name (for error message).

deleteMap :: forall k v s. (MapInstrs map, IsComparable k, KnownValue k, KnownValue v) => (k ': (map k v ': s)) :-> (map k v ': s)

Delete element from the map.

setDelete :: IsComparable e => (e & (Set e & s)) :-> (Set e & s)

Delete given element from the set.

Additional Morley macros

data View (a :: Type) (r :: Type)

view type synonym as described in A1.

Constructors

View

Fields

viewParam :: a

viewCallbackTo :: ContractAddr r

data Void_ (a :: Type) (b :: Type)

void type synonym as described in A1.

Constructors

Void_

Fields

voidParam :: a

Entry point argument.

voidResProxy :: Lambda b b

Type of result reported via failWith.

newtype VoidResult r

Newtype over void result type used in tests to distinguish successful void result from other errors.

Usage example: lExpectFailWith (== VoidResult roleMaster)`

Constructors

VoidResult

Fields

unVoidResult :: r

view_ :: (KnownValue r, NoOperation r, NoBigMap r) => (forall s0. ((a, storage) & s0) :-> (r ': s0)) -> (View a r & (storage & s)) :-> ((List Operation, storage) & s)

void_ :: forall a b s s' anything. (Typeable b, IsoValue b, KnownValue b) => ((a & s) :-> (b & s')) -> (Void_ a b & s) :-> anything

mkVoid :: forall b a. a -> Void_ a b