Lorentz.Rebinded

Description

Reimplementation of some syntax sugar.

You need the following module pragmas to make it work smoothly:

Documentation

(>>) :: (a :-> b) -> (b :-> c) -> a :-> c

Aliases for '(#)' used by do-blocks.

pure :: Applicative f => a -> f a

Lift a value.

return :: Monad m => a -> m a

Inject a value into the monadic type.

ifThenElse :: Condition st arg argl argr -> (argl :-> o) -> (argr :-> o) -> arg :-> o

Defines semantics of if ... then ... else ... construction.

data Condition st arg argl argr where

Predicate for if ... then .. else ... construction, defines a kind of operation applied to the top elements of the current stack.

Constructors

Holds :: Condition s (Bool ': s) s s

IsSome :: Condition s (Maybe a ': s) (a ': s) s

IsNone :: Condition s (Maybe a ': s) s (a ': s)

IsLeft :: Condition s (Either l r ': s) (l ': s) (r ': s)

IsRight :: Condition s (Either l r ': s) (r ': s) (l ': s)

IsCons :: Condition s ([a] ': s) (a ': ([a] ': s)) s

IsNil :: Condition s ([a] ': s) s (a ': ([a] ': s))

IsZero :: (UnaryArithOpHs Eq' a, UnaryArithResHs Eq' a ~ Bool) => Condition s (a ': s) s s

IsNotZero :: (UnaryArithOpHs Eq' a, UnaryArithResHs Eq' a ~ Bool) => Condition s (a ': s) s s

IsEq :: IfCmpXConstraints a Eq' => Condition s (a ': (a ': s)) s s

IsNeq :: IfCmpXConstraints a Neq => Condition s (a ': (a ': s)) s s

IsLt :: IfCmpXConstraints a Lt => Condition s (a ': (a ': s)) s s

IsGt :: IfCmpXConstraints a Gt => Condition s (a ': (a ': s)) s s

IsLe :: IfCmpXConstraints a Le => Condition s (a ': (a ': s)) s s

IsGe :: IfCmpXConstraints a Ge => Condition s (a ': (a ': s)) s s

Re-exports required for RebindableSyntax

fromInteger :: Num a => Integer -> a

Conversion from an Integer. An integer literal represents the application of the function fromInteger to the appropriate value of type Integer, so such literals have type (Num a) => a.

fromString :: IsString a => String -> a

fromLabel :: IsLabel x a => a