Lorentz.UStore.Haskell

Description

Conversion between UStore in Haskell and Michelson representation.

Documentation

type UStoreContent = [(ByteString, ByteString)]

UStore content represented as key-value pairs.

class GUStoreConversible (Rep template) => UStoreConversible template

Given template can be converted to UStore value.

mkUStore :: (Generic template, UStoreConversible template) => template -> UStore template

Make UStore from separate big_maps and fields.

ustoreDecompose :: forall template. (Generic template, UStoreConversible template) => UnpackEnv -> UStore template -> Either Text (UStoreContent, template)

Decompose UStore into separate big_maps and fields.

Since this function needs to UNPACK content of UStore to actual keys and values, you have to provide UnpackEnv.

Along with resulting value, you get a list of UStore entries which were not recognized as belonging to any submap or field according to UStore's template - this should be empty unless UStore invariants were violated.

ustoreDecomposeFull :: forall template. (Generic template, UStoreConversible template) => UnpackEnv -> UStore template -> Either Text template

Like ustoreDecompose, but requires all entries from UStore to be recognized.