Lorentz.UStore.Types

Description

UStore definition and common type-level stuff.

UStore and related type definitions

newtype UStore a

Gathers multple fields and BigMaps under one object.

Type argument of this datatype stands for a "store template" - a datatype with one constructor and multiple fields, each containing an object of type UStoreField or |~> and corresponding to single virtual field or BigMap respectively. It's also possible to parameterize it with a larger type which is a product of types satisfying the above property.

Inside it keeps only one BigMap thus not violating Michelson limitations.

Constructors

UStore

Fields

unUStore :: BigMap ByteString ByteString

newtype k |~> v

Describes one virtual big map in the storage.

Constructors

UStoreSubMap

Fields

unUStoreSubMap :: Map k v

newtype UStoreField v

Constructors

UStoreField

Fields

unUStoreField :: v

Type-lookup-by-name

type GetUStoreKey store name = MSKey (GetUStore name store)

Get type of submap key.

type GetUStoreValue store name = MSValue (GetUStore name store)

Get type of submap value.

type GetUStoreField store name = FSValue (GetUStore name store)

Get type of plain field.

Internals

data ElemSignature

What was found on lookup by constructor name.

This keeps either type arguments of |~> or UStoreField.

Constructors

MapSignature Type Type

FieldSignature Type

type GetUStore name a = MERequireFound name a (GLookupStore name (Rep a))

Get map signature from the constructor with a given name.

type family MSKey ms where ...

Equations

MSKey (MapSignature k _) = k

MSKey (FieldSignature _) = TypeError (Text "Expected UStore submap, but field was referred")

type family MSValue ms where ...

Equations

MSValue (MapSignature _ v) = v

MSValue (FieldSignature _) = TypeError (Text "Expected UStore submap, but field was referred")

type family FSValue ms where ...

Equations

FSValue (FieldSignature v) = v

FSValue (MapSignature _ _) = TypeError (Text "Expected UStore field, but submap was referred")