Lorentz.Value

Description

Re-exports typed Value, CValue, some core types, some helpers and defines aliases for constructors of typed values.

Documentation

type Value = Value' Instr

class IsoValue a where

Isomorphism between Michelson values and plain Haskell types.

Default implementation of this typeclass converts ADTs to Michelson "pair"s and "or"s.

Minimal complete definition

Nothing

Associated Types

type ToT a :: T

Type function that converts a regular Haskell type into a T type.

Methods

toVal :: a -> Value (ToT a)

Converts a Haskell structure into Value representation.

toVal :: (Generic a, GIsoValue (Rep a), ToT a ~ GValueType (Rep a)) => a -> Value (ToT a)

Converts a Haskell structure into Value representation.

fromVal :: Value (ToT a) -> a

Converts a Value into Haskell type.

fromVal :: (Generic a, GIsoValue (Rep a), ToT a ~ GValueType (Rep a)) => Value (ToT a) -> a

Converts a Value into Haskell type.

class IsoCValue a where

Isomorphism between Michelson primitive values and plain Haskell types.

Associated Types

type ToCT a :: CT

Type function that converts a regular Haskell type into a comparable type (which has kind CT).

Methods

toCVal :: a -> CValue (ToCT a)

Converts a single Haskell value into CVal representation.

fromCVal :: CValue (ToCT a) -> a

Converts a CVal value into a single Haskell value.

data CValue t where

Representation of comparable value in Michelson language.

By specification, we're allowed to compare only following types: int, nat, string, bytes, mutez, bool, key_hash, timestamp, address.

Only these values can be used as map keys or set elements.

Constructors

CvInt :: Integer -> CValue CInt

CvNat :: Natural -> CValue CNat

CvString :: MText -> CValue CString

CvBytes :: ByteString -> CValue CBytes

CvMutez :: Mutez -> CValue CMutez

CvBool :: Bool -> CValue CBool

CvKeyHash :: KeyHash -> CValue CKeyHash

CvTimestamp :: Timestamp -> CValue CTimestamp

CvAddress :: Address -> CValue CAddress

Primitive types

data Integer

Invariant: Jn# and Jp# are used iff value doesn't fit in S#

Useful properties resulting from the invariants:

abs (S# _) <= abs (Jp# _)

abs (S# _) <  abs (Jn# _)

data Natural

Type representing arbitrary-precision non-negative integers.

>>> 2^100 :: Natural
1267650600228229401496703205376

Operations whose result would be negative throw (Underflow :: ArithException),

>>> -1 :: Natural
*** Exception: arithmetic underflow

Since: base-4.8.0.0

data MText

Michelson string value.

This is basically a mere text with limits imposed by the language: http://tezos.gitlab.io/zeronet/whitedoc/michelson.html#constants Although, this document seems to be not fully correct, and thus we applied constraints deduced empirically.

You construct an item of this type using one of the following ways:

With QuasyQuotes when need to create a string literal.

>>> [mt|Some text|]
MTextUnsafe { unMText = "Some text" }

With mkMText when constructing from a runtime text value.

With mkMTextUnsafe or MTextUnsafe when absolutelly sure that given string does not violate invariants.

With mkMTextCut when not sure about text contents and want to make it compliant with Michelson constraints.

data Bool

Constructors

False

True

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Char8 it can be interpreted as containing 8-bit characters.

data Address

Data type corresponding to address structure in Tezos.

data Mutez

Mutez is a wrapper over integer data type. 1 mutez is 1 token (μTz).

data Timestamp

Time in the real world. Use the functions below to convert it to/from Unix time in seconds.

data KeyHash

b58check of a public key.

data PublicKey

ED25519 public cryptographic key.

data Signature

ED25519 cryptographic signature.

data Set a

A set of values a.

data Map k a

A Map from keys k to values a.

newtype BigMap k v

Constructors

BigMap

Fields

unBigMap :: Map k v

type Operation = Operation' Instr

data Maybe a

The Maybe type encapsulates an optional value. A value of type Maybe a either contains a value of type a (represented as Just a), or it is empty (represented as Nothing). Using Maybe is a good way to deal with errors or exceptional cases without resorting to drastic measures such as error.

The Maybe type is also a monad. It is a simple kind of error monad, where all errors are represented by Nothing. A richer error monad can be built using the Either type.

Constructors

Nothing

Just a

type List = []

newtype ContractAddr (cp :: Type)

Since Contract name is used to designate contract code, lets call analogy of TContract type as follows.

Constructors

ContractAddr

Fields

unContractAddress :: Address

Constructors

toMutez :: Word32 -> Mutez

Safely create Mutez.

This is recommended way to create Mutez from a numeric literal; you can't construct all valid Mutez values using this function but for small values it works neat.

Warnings displayed when trying to construct invalid Natural or Word literal are hardcoded for these types in GHC implementation, so we can only exploit these existing rules.

mt :: QuasiQuoter

QuasyQuoter for constructing Michelson strings.

Validity of result will be checked at compile time. Note:

slash must be escaped

newline character must appear as '\n'

use quotes as is

other special characters are not allowed.

timestampFromSeconds :: Integer -> Timestamp

timestampFromUTCTime :: UTCTime -> Timestamp

timestampQuote :: QuasiQuoter

Quote a value of type Timestamp in yyyy-mm-ddThh:mm:ss[.sss]Z format.

>>> formatTimestamp [timestampQuote| 2019-02-21T16:54:12.2344523Z |]
"2019-02-21T16:54:12Z"

Inspired by 'time-quote' library.

Misc

class Default a where

A class for types with a default value.

Minimal complete definition

Nothing

Methods

def :: a

The default value for this type.