Michelson.Doc

Description

Renderable documentation injected to contract code.

Documentation

class (Typeable d, DOrd d, KnownNat (DocItemPosition d)) => DocItem d where

A piece of documentation describing one property of a thing, be it a name or description of a contract, or an error throwable by given endpoint.

Items of the same type appear close to each other in a rendered documentation and form a section.

Doc items are later injected into a contract code via a dedicated nop-like instruction. Normally doc items which belong to one section appear in resulting doc in the same order in which they appeared in the contract.

While documentation framework grows, this typeclass acquires more and more methods for fine tuning of existing rendering logic because we don't want to break backward compatibility, hope one day we will make everything concise :( E.g. all rendering and reording stuff could be merged in one method, and we could have several template implementations for it which would allow user to specify only stuff relevant to his case.

Minimal complete definition

docItemSectionName, docItemToMarkdown

Associated Types

type DocItemPosition d = (pos :: Nat) | pos -> d

Position of this item in the resulting documentation; the smaller the value, the higher the section with this element will be placed.

Documentation structure is not necessarily flat. If some doc item consolidates a whole documentation block within it, this block will have its own placement of items independent from outer parts of the doc.

type DocItemPlacement d :: DocItemPlacementKind

Defines where given doc item should be put. There are two options: 1. Inline right here (default behaviour); 2. Put into definitions section.

Note that we require all doc items with "in definitions" placement to have Eq and Ord instances which comply the following law: if two documentation items describe the same entity or property, they should be considered equal.

Methods

docItemSectionName :: Maybe Text

When multiple items of the same type belong to one section, how this section will be called.

If not provided, section will contain just untitled content.

docItemSectionDescription :: Maybe Markdown

Description of a section.

Can be used to mention some common things about all elements of this section. Markdown syntax is permitted here.

docItemSectionNameStyle :: DocSectionNameStyle

How to render section name.

Takes effect only if section name is set.

docItemRef :: d -> DocItemRef (DocItemPlacement d)

Defines a function which constructs an unique identifier of given doc item, if it has been decided to put the doc item into definitions section.

Identifier should be unique both among doc items of the same type and items of other types. Thus, consider using "typeId-contentId" pattern.

docItemRef :: DocItemPlacement d ~ DocItemInlined => d -> DocItemRef (DocItemPlacement d)

Defines a function which constructs an unique identifier of given doc item, if it has been decided to put the doc item into definitions section.

Identifier should be unique both among doc items of the same type and items of other types. Thus, consider using "typeId-contentId" pattern.

docItemToMarkdown :: HeaderLevel -> d -> Markdown

Render given doc item to Markdown, preferably one line, optionally with header.

Accepts the smallest allowed level of header. (Using smaller value than provided one will interfere with existing headers thus delivering mess).

docItemDependencies :: d -> [SomeDocDefinitionItem]

All doc items which this doc item refers to.

They will automatically be put to definitions as soon as given doc item is detected.

docItemsOrder :: [d] -> [d]

This function accepts doc items put under the same section in the order in which they appeared in the contract and returns their new desired order. It's also fine to use this function for filtering or merging doc items.

Default implementation * leaves inlined items as is; * for items put to definitions, lexicographically sorts them by their id.

docItemPosition :: forall d. DocItem d => DocItemPos

Get doc item position at term-level.

newtype DocItemId

Some unique identifier of a doc item.

All doc items which should be refer-able need to have this identifier.

Constructors

DocItemId Text

data DocItemPlacementKind

Where do we place given doc item.

Constructors

DocItemInlined
Placed in the document content itself.

DocItemInDefinitions
Placed in dedicated definitions section; can later be referenced.

data DocItemRef (p :: DocItemPlacementKind) where

Defines an identifier which given doc item can be referenced with.

Constructors

DocItemRef :: DocItemId -> DocItemRef DocItemInDefinitions

DocItemNoRef :: DocItemRef DocItemInlined

data DocSectionNameStyle

How to render section name.

Constructors

DocSectionNameBig
Suitable for block name.

DocSectionNameSmall
Suitable for subsection title within block.

data SomeDocItem where

Hides some documentation item.

Constructors

SomeDocItem :: DocItem d => d -> SomeDocItem

data SomeDocDefinitionItem where

Hides some documentation item which is put to "definitions" section.

Constructors

SomeDocDefinitionItem :: (DocItem d, DocItemPlacement d ~ DocItemInDefinitions) => d -> SomeDocDefinitionItem

data DocElem d

A doc item which we store, along with related information.

Constructors

DocElem

Fields

deItem :: d

Doc item itself.

deSub :: Maybe SubDoc

Subdocumentation, if given item is a group.

data DocSection

Several doc items of the same type.

Constructors

DocItem d => DocSection (NonEmpty $ DocElem d)

type DocBlock = Map DocItemPos DocSection

A map from positions to document elements.

This form effeciently keeps documentation for its incremental building. Doc items here appear close to how they were located in the contract; for instance, docItemsOrder is not yet applied at this stage. You only can be sure that items within each group are splitted across sections correctly.

newtype SubDoc

A part of documentation to be grouped. Essentially incapsulates DocBlock.

Constructors

SubDoc DocBlock

data ContractDoc

Keeps documentation gathered for some piece of contract code.

Used for building documentation of a contract.

Constructors

ContractDoc

Fields

cdContents :: DocBlock

All inlined doc items.

cdDefinitions :: DocBlock

Definitions used in document.

Usually you put some large and repetitive descriptions here. This differs from the document content in that it contains sections which are always at top-level, disregard the nesting.

All doc items which define docItemId method go here, and only they.

cdDefinitionsSet :: Set SomeDocDefinitionItem

We remember all already declared entries to avoid cyclic dependencies in documentation items discovery.

cdDefinitionIds :: Set DocItemId

We remember all already used identifiers. (Documentation naturally should not declare multiple items with the same identifier because that would make references to the respective anchors ambiguous).

type DocGrouping = SubDoc -> SomeDocItem

A function which groups a piece of doc under one doc item.

cdContentsL :: Lens' ContractDoc DocBlock

cdDefinitionsL :: Lens' ContractDoc DocBlock

cdDefinitionsSetL :: Lens' ContractDoc (Set SomeDocDefinitionItem)

cdDefinitionIdsL :: Lens' ContractDoc (Set DocItemId)

deIsAtomic :: DocElem d -> Bool

Whether given DocElem is atomic.

Normally, atomic DocElems are ones appearing in DOC_ITEM instruction, and non-atomic ones are put to DocGroup.

subDocToMarkdown :: HeaderLevel -> SubDoc -> Markdown

Render documentation for SubDoc.

docItemToBlock :: forall di. DocItem di => di -> DocBlock

Lift an atomic doc item to a block.

lookupDocBlockSection :: forall d. DocItem d => DocBlock -> Maybe (NonEmpty d)

Find all doc items of the given type.

contractDocToMarkdown :: ContractDoc -> LText

Render given contract documentation to markdown document.

docGroupContent :: DocGrouping -> ContractDoc -> ContractDoc

Apply given grouping to documentation being built.

docDefinitionRef :: (DocItem d, DocItemPlacement d ~ DocItemInDefinitions) => Markdown -> d -> Markdown

Make a reference to doc item in definitions.

newtype DGeneralInfoSection

General (meta-)information about the contract such as git revision, contract's authors, etc. Should be relatively short (not several pages) because it is put somewhere close to the beginning of documentation.

Constructors

DGeneralInfoSection SubDoc

data DName

Give a name to document block.

Constructors

DName Text SubDoc

data DDescription

Description of something.

Constructors

DDescription Markdown

data DGitRevision

Constructors

DGitRevisionKnown DGitRevisionInfo

DGitRevisionUnknown

newtype GitRepoSettings

Repository settings for DGitRevision.

Constructors

GitRepoSettings

Fields

grsMkGitRevision :: Text -> Text

By commit sha make up a url to that commit in remote repository.

mkDGitRevision :: ExpQ

Make DGitRevision.

>>> :t $mkDGitRevision
GitRepoSettings -> DGitRevision

morleyRepoSettings :: GitRepoSettings

data DComment

Comment in the doc (mostly used for licenses)

Constructors

DComment Text

data DAnchor

A hand-made anchor.

Constructors

DAnchor Anchor