Michelson.FailPattern

Description

Typical usages of FAILWITH instruction.

Documentation

data TypicalFailWith

This data type captures typical ways to use FAILWITH instruction. Each constructor corresponds to a usage pattern.

typicalFailWithTag :: TypicalFailWith -> MText

Extract error tag out of TypicalFailWith.

isTypicalFailWith :: Instr inp out -> Maybe TypicalFailWith

Check whether given instruction ends with a typical FAILWITH usage. It does not recursively check instructions that can be passed to other instructions.

modifyTypicalFailWith :: HasCallStack => (MText -> SomeConstrainedValue ConstantScope') -> Instr inp out -> Instr inp out

If given instruction ends with a typical FAILWITH usage, modify the tag used there using given transformation function. It can return any value, not necessarily a string.

class (Typeable a, ConstantScope a) => ConstantScope' a

We need this class to pass it to SomeConstrainedValue. ConstantScope is needed because we push a value and Typeable is needed for FAILWITH.