Michelson.Interpret

Description

Module, containing function to interpret Michelson instructions against given context and input stack.

Documentation

data ContractEnv

Environment for contract execution.

Constructors

ContractEnv

Fields

ceNow :: Timestamp

Timestamp returned by the NOW instruction.

ceMaxSteps :: RemainingSteps

Number of steps after which execution unconditionally terminates.

ceBalance :: Mutez

Current amount of mutez of the current contract.

ceContracts :: TcOriginatedContracts

Mapping from existing contracts' addresses to their executable representation.

ceSelf :: Address

Address of the interpreted contract.

ceSource :: Address

The contract that initiated the current transaction.

ceSender :: Address

The contract that initiated the current internal transaction.

ceAmount :: Mutez

Amount of the current transaction.

ceChainId :: ChainId

Identifier of the current chain.

data InterpreterState

Constructors

InterpreterState

Fields

isMorleyLogs :: MorleyLogs

isRemainingSteps :: RemainingSteps

data MichelsonFailed where

Represents `[FAILED]` state of a Michelson program. Contains value that was on top of the stack when FAILWITH was called.

Constructors

MichelsonFailedWith :: (Typeable t, SingI t) => Value t -> MichelsonFailed

MichelsonArithError :: (Typeable n, Typeable m) => ArithError (CValue n) (CValue m) -> MichelsonFailed

MichelsonGasExhaustion :: MichelsonFailed

MichelsonFailedTestAssert :: Text -> MichelsonFailed

MichelsonAmbigousEpRef :: EpName -> EpAddress -> MichelsonFailed

newtype RemainingSteps

Constructors

RemainingSteps Word64

data SomeItStack where

Constructors

SomeItStack :: ExtInstr inp -> Rec Value inp -> SomeItStack

newtype MorleyLogs

Morley logs for interpreter state.

Constructors

MorleyLogs

Fields

unMorleyLogs :: [Text]

Logs in reverse order.

noMorleyLogs :: MorleyLogs

interpret :: ContractCode cp st -> EntryPointCallT cp arg -> Value arg -> Value st -> ContractEnv -> ContractReturn st

interpretInstr :: ContractEnv -> Instr inp out -> Rec Value inp -> Either MichelsonFailed (Rec Value out)

Interpret an instruction in vacuum, putting no extra contraints on its execution.

Mostly for testing purposes.

type ContractReturn st = (Either MichelsonFailed ([Operation], Value st), InterpreterState)

mkInitStack :: Value param -> Value st -> Rec Value (ContractInp param st)

fromFinalStack :: Rec Value (ContractOut st) -> ([Operation], Value st)

interpretUntyped :: Contract -> Value -> Value -> ContractEnv -> Either InterpretError InterpretResult

Interpret a contract without performing any side effects. This function uses untyped representation of contract, parameter and storage. Mostly used for testing.

data InterpretError

Constructors

RuntimeFailure (MichelsonFailed, MorleyLogs)

IllTypedContract TCError

IllTypedParam TCError

IllTypedStorage TCError

data InterpretResult where

Constructors

InterpretResult

Fields

:: StorageScope st

=> { iurOps :: [Operation]

, iurNewStorage :: Value st

, iurNewState :: InterpreterState

} -> InterpretResult

type EvalM m = (MonadReader ContractEnv m, MonadState InterpreterState m, MonadError MichelsonFailed m)

type InstrRunner m = forall inp out. Instr inp out -> Rec Value inp -> m (Rec Value out)

runInstr :: EvalM m => InstrRunner m

Function to change amount of remaining steps stored in State monad

runInstrNoGas :: EvalM m => InstrRunner m

runUnpack :: forall t. UnpackedValScope t => ByteString -> Either UnpackError (Value t)

Unpacks given raw data into a typed value.

Internals

initInterpreterState :: ContractEnv -> InterpreterState

handleContractReturn :: StorageScope st => ContractReturn st -> Either InterpretError InterpretResult

runInstrImpl :: EvalM m => InstrRunner m -> InstrRunner m

Function to interpret Michelson instruction(s) against given stack.