Michelson.Parser

Main parser type

type Parser = ReaderT LetEnv (Parsec CustomParserException Text)

Parsers

program :: Parsec CustomParserException Text (Contract' ParsedOp)

Michelson contract with let definitions

value :: Parser ParsedValue

Errors

data CustomParserException

Constructors

UnknownTypeException

StringLiteralException StringLiteralParserException

OddNumberBytesException

WrongTagArgs Natural Positive

WrongAccessArgs Natural Positive

WrongSetArgs Natural Positive

ExcessFieldAnnotation

data ParseErrorBundle s e

A non-empty collection of ParseErrors equipped with PosState that allows to pretty-print the errors efficiently and correctly.

Since: megaparsec-7.0.0

data ParserException

Constructors

ParserException (ParseErrorBundle Text CustomParserException)

data StringLiteralParserException

Constructors

InvalidEscapeSequence Char

InvalidChar Char

Additional helpers

parseNoEnv :: Parser a -> String -> Text -> Either (ParseErrorBundle Text CustomParserException) a

Parse with empty environment

parseValue :: Text -> Either ParserException ParsedValue

Parse untyped value from text which comes from something that is not a file (which is often the case). So we assume it does not need any parsing environment.

parseExpandValue :: Text -> Either ParserException Value

Like parseValue, but also expands macros.

For tests

codeEntry :: Parser [ParsedOp]

Parses code block after "code" keyword of a contract.

This function is part of the module API, its semantics should not change.

type_ :: Parser Type

letType :: Parser LetType

stringLiteral :: Parser ParsedValue

bytesLiteral :: Parser (Value' op)

intLiteral :: Parser (Value' op)

printComment :: Parser PrintComment

Quoters

utypeQ :: QuasiQuoter

Creates Type by its Morley representation.

>>> [utypeQ| (int :a | nat :b) |]
Type (TOr % % (Type (Tc CInt) :a) (Type (Tc CNat) :b)) :

Re-exports

errorBundlePretty

Arguments

:: (Stream s, ShowErrorComponent e)

=> ParseErrorBundle s e
Parse error bundle to display

-> String
Textual rendition of the bundle

Pretty-print a ParseErrorBundle. All ParseErrors in the bundle will be pretty-printed in order together with the corresponding offending lines by doing a single efficient pass over the input stream. The rendered String always ends with a newline.

Since: megaparsec-7.0.0