Michelson.Parser.Helpers

Documentation

mkParser :: (a -> Text) -> a -> Parser a

Make a parser from a string

sepEndBy1 :: MonadPlus m => m a -> m sep -> m (NonEmpty a)

sepBy2 :: MonadPlus m => m a -> m sep -> m (NonEmpty a)

endBy2 p sep parses two or more occurrences of p, separated by sep.

parseDef :: Default a => Parser a -> Parser a

Apply given parser and return default value if it fails.

positive :: Parser Positive

Parse a positive number.

mparens :: Parser a -> Parser a

Parse expression which can be wrapped in parentheses.