Michelson.Preprocess

Description

Apply some transformations to Michelson code.

Documentation

transformStrings :: Bool -> (MText -> MText) -> Instr inp out -> Instr inp out

Transform all strings in a typed instructions using given function. The first argument specifies whether we should go into arguments that contain instructions.

transformBytes :: Bool -> (ByteString -> ByteString) -> Instr inp out -> Instr inp out

Similar to transformStrings but for bytes. TODO [TM-375]: deduplicate