Michelson.Printer

Documentation

class RenderDoc a where

Generalize converting a type into a Text.PrettyPrint.Leijen.Text.Doc. Used to pretty print Michelson code and define Fmt.Buildable instances.

Minimal complete definition

renderDoc

Methods

renderDoc :: RenderContext -> a -> Doc

isRenderable :: a -> Bool

Whether a value can be represented in Michelson code. Normally either all values of some type are renderable or not renderable. However, in case of instructions we have extra instructions which should not be rendered. Note: it's not suficcient to just return mempty for such instructions, because sometimes we want to print lists of instructions and we need to ignore them complete (to avoid putting redundant separators).

printDoc :: Bool -> Doc -> Text

Convert Doc to Text with a line width of 80.

printUntypedContract :: RenderDoc op => Bool -> Contract' op -> Text

Convert an untyped contract into a textual representation which will be accepted by the OCaml reference client.

printTypedContractCode :: (SingI p, SingI s) => Bool -> ContractCode p s -> Text

Convert a typed contract into a textual representation which will be accepted by the OCaml reference client.

printTypedFullContract :: Bool -> FullContract p s -> Text

printSomeContract :: Bool -> SomeContract -> Text

printTypedValue :: forall t. ProperPrintedValBetterErrors t => Bool -> Value t -> Text