Michelson.Printer.Util

Documentation

class RenderDoc a where

Generalize converting a type into a Text.PrettyPrint.Leijen.Text.Doc. Used to pretty print Michelson code and define Fmt.Buildable instances.

Minimal complete definition

renderDoc

Methods

renderDoc :: RenderContext -> a -> Doc

isRenderable :: a -> Bool

Whether a value can be represented in Michelson code. Normally either all values of some type are renderable or not renderable. However, in case of instructions we have extra instructions which should not be rendered. Note: it's not suficcient to just return mempty for such instructions, because sometimes we want to print lists of instructions and we need to ignore them complete (to avoid putting redundant separators).

newtype Prettier a

A new type that can wrap values so that the RenderDoc instances of the combined value can have a different behavior for the pretty printer.

Constructors

Prettier a

printDoc :: Bool -> Doc -> Text

Convert Doc to Text with a line width of 80.

printDocB :: Bool -> Doc -> Builder

Convert Doc to Builder in the same maner as printDoc.

printDocS :: Bool -> Doc -> String

Convert Doc to String in the same maner as printDoc.

renderOps :: RenderDoc op => Bool -> NonEmpty op -> Doc

Generic way to render the different op types that get passed to a contract.

renderOpsList :: RenderDoc op => Bool -> [op] -> Doc

spaces :: Int -> Doc

Create a specific number of spaces.

wrapInParens :: RenderContext -> NonEmpty Doc -> Doc

Wrap documents in parentheses if there are two or more in the list.

buildRenderDoc :: RenderDoc a => a -> Builder

Turn something that is instance of RenderDoc into a Builder. It's formatted the same way as printDoc formats docs.

Smart parentheses

data RenderContext

Environment carried during recursive rendering.

needsParens :: RenderContext

ParensNeeded constant.

doesntNeedParens :: RenderContext

ParensNeeded constant.

addParens :: RenderContext -> Doc -> Doc

Add parentheses if needed.

assertParensNotNeeded :: HasCallStack => RenderContext -> a -> a

Ensure parentheses are not required, for case when you cannot sensibly wrap your expression into them.