Michelson.Runtime.GState

Description

Global blockchain state (emulated).

Auxiliary types

data ContractState

State of a contract with code.

Constructors

ContractState

Fields

csBalance :: Mutez

Amount of mutez owned by this contract.

csStorage :: Value

Storage value associated with this contract.

csContract :: Contract

Contract itself (untyped).

csTypedContract :: Maybe SomeContract

csTypedStorage :: Maybe SomeValue

We keep typed representation of contract code and storage in form, that hides their actual type in order to simplify the rest of the code (e.g. avoid type parameters for ContractState and so on). They are made optional in order to perform safe parsing from JSON (we simply return Nothing in this parser and use getTypedStorage or getTypedContract that optionally typecheck storage or contract code).

getTypedContract :: GState -> ContractState -> Either TCError SomeContract

getTypedStorage :: GState -> ContractState -> Either TCError SomeValue

data SomeContractAndStorage

Constructors

(ParameterScope cp, StorageScope st) => SomeContractAndStorage

Fields

scsContract :: FullContract cp st

scsStorage :: Value st

getTypedContractAndStorage :: (TCError -> err) -> (TCError -> err) -> GState -> ContractState -> Either err SomeContractAndStorage

data AddressState

State of an arbitrary address.

Constructors

ASSimple Mutez
For contracts without code we store only its balance.

ASContract ContractState
For contracts with code we store more state represented by ContractState.

asBalance :: AddressState -> Mutez

Extract balance from AddressState.

GState

data GState

Persistent data passed to Morley contracts which can be updated as result of contract execution.

Constructors

GState

Fields

gsChainId :: ChainId

Identifier of chain.

gsAddresses :: Map Address AddressState

All known addresses and their state.

gsChainIdL :: Lens' GState ChainId

gsAddressesL :: Lens' GState (Map Address AddressState)

genesisAddresses :: NonEmpty Address

Initially these addresses have a lot of money.

genesisKeyHashes :: NonEmpty KeyHash

KeyHash of genesis address.

genesisAddress :: Address

One of genesis addresses.

More genesisAddresses which can be used in tests

genesisAddress1 :: Address

More genesis addresses

We know size of genesisAddresses, so it is safe to use !!

genesisAddress2 :: Address

More genesis addresses

We know size of genesisAddresses, so it is safe to use !!

genesisAddress3 :: Address

More genesis addresses

We know size of genesisAddresses, so it is safe to use !!

genesisAddress4 :: Address

genesisAddress5 :: Address

genesisAddress6 :: Address

genesisKeyHash :: KeyHash

One of genesis key hashes.

initGState :: GState

Initial GState. It's supposed to be used if no GState is provided. It puts plenty of money on each genesis address.

readGState :: FilePath -> IO GState

Read GState from a file.

writeGState :: FilePath -> GState -> IO ()

Write GState to a file.

Operations on GState

data GStateUpdate

Updates that can be applied to GState.

Constructors

GSAddAddress Address AddressState

GSSetStorageValue Address Value SomeValue

GSSetBalance Address Mutez

data GStateUpdateError

Constructors

GStateAddressExists Address

GStateUnknownAddress Address

GStateNotContract Address

applyUpdate :: GStateUpdate -> GState -> Either GStateUpdateError GState

Apply GStateUpdate to GState.

applyUpdates :: [GStateUpdate] -> GState -> Either GStateUpdateError GState

Apply a list of GStateUpdates to GState.

extractAllContracts :: GState -> TcOriginatedContracts

Retrive all contracts stored in GState