Michelson.Test.Dummy

Description

Dummy data to be used in tests where it's not essential.

Documentation

dummyNow :: Timestamp

Dummy timestamp, can be used to specify current NOW value or maybe something else.

dummyMaxSteps :: RemainingSteps

Dummy value for maximal number of steps a contract can make. Intentionally quite large, because most likely if you use dummy value you don't want the interpreter to stop due to gas exhaustion. On the other hand, it probably still prevents the interpreter from working for eternity.

dummyContractEnv :: ContractEnv

Dummy ContractEnv with some reasonable hardcoded values. You can override values you are interested in using record update syntax.

dummyOrigination :: Value -> Contract -> OriginationOperation

OriginationOperation with most data hardcoded to some reasonable values. Contract and initial values must be passed explicitly, because otherwise it hardly makes sense.