Michelson.Test.Integrational

Description

Utilities for integrational testing. Example tests can be found in the 'morley-test' test suite.

Re-exports

data TxData

Data associated with a particular transaction.

Constructors

TxData

Fields

tdSenderAddress :: Address

tdParameter :: Value

tdEntrypoint :: EpName

tdAmount :: Mutez

More genesis addresses which can be used in tests

genesisAddress :: Address

One of genesis addresses.

genesisAddress1 :: Address

More genesis addresses

We know size of genesisAddresses, so it is safe to use !!

genesisAddress2 :: Address

More genesis addresses

We know size of genesisAddresses, so it is safe to use !!

genesisAddress3 :: Address

More genesis addresses

We know size of genesisAddresses, so it is safe to use !!

genesisAddress4 :: Address

genesisAddress5 :: Address

genesisAddress6 :: Address

Testing engine

type IntegrationalValidator = Either (ExecutorError -> Bool) SuccessValidator

Validator for integrational testing. If an error is expected, it should be Left with validator for errors. Otherwise it should check final global state and its updates.

type SuccessValidator = InternalState -> GState -> [GStateUpdate] -> Either ValidationError ()

Validator for integrational testing that expects successful execution.

type IntegrationalScenarioM = StateT InternalState (Except ScenarioError)

A monad inside which integrational tests can be described using do-notation.

type IntegrationalScenario = IntegrationalScenarioM Validated

data ValidationError

Constructors

UnexpectedExecutorError IntegrationalExecutorError

UnexpectedTypeCheckError TCError

ExpectingInterpreterToFail

IncorrectUpdates ValidationError [GStateUpdate]

IncorrectStorageUpdate AddressName Text

InvalidStorage AddressName ExpectedStorage Text

StoragePredicateMismatch AddressName Text

InvalidBalance AddressName ExpectedBalance Text

UnexpectedUpdates (NonEmpty GStateUpdate)

CustomValidationError Text

integrationalTestExpectation :: HasCallStack => IntegrationalScenario -> Expectation

Integrational test that executes given operations and validates them using given validator. It can fail using Expectation capability. It starts with initGState and some reasonable dummy values for gas limit and current timestamp. You can update blockchain state by performing some operations.

integrationalTestProperty :: IntegrationalScenario -> Property

Integrational test similar to integrationalTestExpectation. It can fail using Property capability. It can be used with QuickCheck's forAll to make a property-based test with arbitrary data.

originate :: Contract -> Text -> Value -> Mutez -> IntegrationalScenarioM Address

Originate a contract with given initial storage and balance. Its address is returned.

tOriginate :: (ParameterScope cp, StorageScope st) => FullContract cp st -> Text -> Value st -> Mutez -> IntegrationalScenarioM Address

Like originate, but for typed contract and value.

transfer :: TxData -> Address -> IntegrationalScenarioM ()

Transfer tokens to a given address.

tTransfer :: forall arg. ParameterScope arg => ("from" :! Address) -> ("to" :! Address) -> Mutez -> EpName -> Value arg -> IntegrationalScenarioM ()

Similar to transfer, for typed values. Note that it works with untyped Address and does not check that entrypoint with given name is present and has the expected type. Passed value must correspond to the entrypoint argument type, not the parameter type of the contract (and must be unit for implicit accounts).

validate :: IntegrationalValidator -> IntegrationalScenario

Validate the execution result.

integrationalFail :: ValidationError -> IntegrationalScenarioM anything

Just fail with given error.

setMaxSteps :: RemainingSteps -> IntegrationalScenarioM ()

Make all further interpreter calls (which are triggered by the validate function) use given gas limit.

modifyNow :: (Timestamp -> Timestamp) -> IntegrationalScenarioM ()

Make all further interpreter calls (which are triggered by the validate function) use modified timestamp as the current one.

setNow :: Timestamp -> IntegrationalScenarioM ()

Make all further interpreter calls (which are triggered by the validate function) use given timestamp as the current one.

rewindTime :: Integer -> IntegrationalScenarioM ()

Increase current time by the given number of seconds.

withSender :: Address -> IntegrationalScenarioM a -> IntegrationalScenarioM a

Pretend that given address initiates all the transfers within the code block (i.e. SENDER instruction will return this address).

setChainId :: ChainId -> IntegrationalScenarioM ()

Make all further interpreter calls (which are triggered by the validate function) use given chain id.

branchout :: HasCallStack => [(Text, IntegrationalScenario)] -> IntegrationalScenario

Execute multiple testing scenarios independently, basing them on scenario built till this point.

The following property holds for this function:

pre >> branchout [a, b, c] = branchout [pre >> a, pre >> b, pre >> c] .

In case of property failure in one of the branches no following branch is executed.

Providing empty list of scenarios to this function causes error; we do not require NonEmpty here though for convenience.

(?-) :: Text -> a -> (Text, a) infixr 0

Make a tuple with name without extra syntactic noise.

offshoot :: Text -> IntegrationalScenario -> IntegrationalScenarioM ()

Test given scenario with the state gathered till this moment; if this scenario passes, go on as if it never happened.

Validators

composeValidators :: SuccessValidator -> SuccessValidator -> SuccessValidator

Compose two success validators.

For example:

expectBalance bal addr composeValidators expectStorageUpdateConst addr2 ValueUnit

composeValidatorsList :: [SuccessValidator] -> SuccessValidator

Compose a list of success validators.

expectAnySuccess :: SuccessValidator

SuccessValidator that always passes.

expectNoUpdates :: SuccessValidator

Check that there were no updates.

expectNoStorageUpdates :: SuccessValidator

Check that there were no storage updates.

expectStorageUpdate :: Address -> (Value -> Either ValidationError ()) -> SuccessValidator

Check that storage value is updated for given address. Takes a predicate that is used to check the value.

It works even if updates are not filtered (i. e. a value can be updated more than once).

expectStorageUpdateConst :: Address -> Value -> SuccessValidator

Like expectStorageUpdate, but expects a constant.

expectBalance :: Address -> Mutez -> SuccessValidator

Check that eventually address has some particular balance.

expectStorage :: Address -> (Value -> Either ValidationError ()) -> SuccessValidator

Check that storage value satisfies the given predicate.

expectStorageConst :: Address -> Value -> SuccessValidator

Check that eventually address has some particular storage value.

tExpectStorageConst :: forall st. StorageScope st => Address -> Value st -> SuccessValidator

Similar to expectStorageConst, for typed stuff.

expectGasExhaustion :: ExecutorError -> Bool

Check that interpreter failed due to gas exhaustion.

expectMichelsonFailed :: (MichelsonFailed -> Bool) -> Address -> ExecutorError -> Bool

Expect that interpretation of contract with given address ended with [FAILED].