Michelson.Test.Util

Description

Testing utility functions used by testing framework itself or intended to be used by test writers.

Documentation

leftToShowPanic :: (Show e, HasCallStack) => Either e a -> a

leftToPrettyPanic :: (Buildable e, HasCallStack) => Either e a -> a

failedProp :: Text -> Property

A Property that always failes with given message.

succeededProp :: Property

A Property that always succeeds.

qcIsLeft :: Show b => Either a b -> Property

The Property holds on `Left a`.

qcIsRight :: Show a => Either a b -> Property

The Property holds on `Right b`.

roundtripTest :: forall x y err. (Show x, Show err, Typeable x, Arbitrary x, Eq x, Eq err) => (x -> y) -> (y -> Either err x) -> TestTree

This TestTree contains a property based test for conversion from some x to some y and back to x (it should successfully return the initial x).