Michelson.TypeCheck

Documentation

typeCheckContract :: TcOriginatedContracts -> Contract -> Either TCError SomeContract

typeCheckTopLevelType :: HasCallStack => TcOriginatedContracts -> Type -> Value -> Either TCError SomeValue

typeVerifyTopLevelType :: (SingI t, HasCallStack) => TcOriginatedContracts -> Value -> Either TCError (Value t)

typeCheckValue :: forall t. SingI t => Value -> TypeCheckInstr (Value t)

Function typeCheckValue converts a single Michelson value given in representation from Michelson.Untyped module hierarchy to representation in strictly typed GADT.

typeCheckValue is polymorphic in the expected type of value.

Type checking algorithm pattern-matches on parse value representation, expected type t and constructs Val t value.

If there was no match on a given pair of value and expected type, that is interpreted as input of wrong type and type check finishes with error.

typeCheckList :: Typeable inp => [ExpandedOp] -> HST inp -> TypeCheck (SomeInstr inp)

Function typeCheckList converts list of Michelson instructions given in representation from Michelson.Type module to representation in strictly typed GADT.

Types are checked along the way which is neccessary to construct a strictly typed value.

As a second argument, typeCheckList accepts input stack type representation.

typeCheckCValue :: forall op ct. Value' op -> Sing ct -> Either (Value' op, TCTypeError) (CValue ct)

typeCheckExt :: forall s. Typeable s => TypeCheckListHandler s -> ExpandedExtInstr -> HST s -> TypeCheckInstr (SomeInstr s)

module Michelson.TypeCheck.Error

module Michelson.TypeCheck.Types

module Michelson.TypeCheck.TypeCheck

eqType :: forall (a :: T) (b :: T). Each [Typeable, SingI] [a, b] => Either TCTypeError (a :~: b)

Function eqType is a simple wrapper around Data.Typeable.eqT suited for use within Either TCTypeError a applicative.

matchTypes :: forall t1 t2. Each [Typeable, SingI] [t1, t2] => Notes t1 -> Notes t2 -> Either TCTypeError (t1 :~: t2, Notes t1)

Check whether given types are structurally equal and annotations converge.