Michelson.TypeCheck.Error

Description

Errors that can occur when some code is being typechecked.

Documentation

data NotEnoughItemsInstr

Description of the instruction which wants more items on stack than currently present.

Constructors

NotEnoughDrop

NotEnoughDip

NotEnoughDig

NotEnoughDug

data TCTypeError

Data type that represents various errors which are related to type system. These errors are used to specify info about type check errors in TCError data type.

Constructors

AnnError AnnConvergeError
Annotation unify error

TypeEqError T T
Type equality error

StackEqError [T] [T]
Stacks equality error

UnsupportedTypes [T]
Error that happens when type cannot be used in the corresponding scope. Argument of this constructor carries types which, in the aggregate, violate the restriction (e.g. timestamp and timestamp passed to MUL instruction).

InvalidValueType T
Error that happens when a Value is never a valid source for this type (e.g. timestamp cannot be obtained from a ValueTrue)

NotEnoughItemsOnStack Word NotEnoughItemsInstr
There are not enough items on stack to perform a certain instruction.

IllegalEntryPoint EpNameFromRefAnnError
Invalid entrypoint name provided

UnknownContract Address
Contract with given address is not originated.

EntryPointNotFound EpName
Given entrypoint is not present.

IllegalParamDecl ParamEpError
Incorrect parameter declaration (with respect to entrypoints feature).

NegativeNat
Natural numbers cannot be negative

MutezOverflow
Exceeds the maximal mutez value

InvalidAddress ParseEpAddressError
Address couldn't be parsed from its textual representation

InvalidKeyHash CryptoParseError
KeyHash couldn't be parsed from its textual representation

InvalidTimestamp
Timestamp is not RFC339 compliant

data TCError

Type check error

Constructors

TCFailedOnInstr ExpandedInstr SomeHST Text InstrCallStack (Maybe TCTypeError)

TCFailedOnValue Value T Text InstrCallStack (Maybe TCTypeError)

TCContractError Text (Maybe TCTypeError)

TCUnreachableCode InstrCallStack (NonEmpty ExpandedOp)

TCExtError SomeHST InstrCallStack ExtError

data ExtError

Various type errors possible when checking Morley extension commands

Constructors

LengthMismatch StackTypePattern

VarError Text StackFn

TypeMismatch StackTypePattern Int TCTypeError

TyVarMismatch Var Type StackTypePattern Int TCTypeError

StkRestMismatch StackTypePattern SomeHST SomeHST TCTypeError

TestAssertError Text

InvalidStackReference StackRef StackSize

newtype StackSize

Constructors

StackSize Natural