Michelson.Typed.Annotation

Description

Module, providing Notes t data type, which holds annotations for a given type t.

Annotation type Notes t is a tree, each leaf is either a star (*) or a constructor holding some annotation data for a given type t. Star corresponds to the case when given Michelson type contains no annotations.

This module also provides type class Converge along with some utility functions which are used to combine two annotations trees a and b into a new one c in such a way that c can be obtained from both a and b by replacing some * leafs with type or/and field annotations.

Documentation

data Notes t where

Data type, holding annotation data for a given Michelson type t.

Each constructor corresponds to exactly one constructor of T and holds all type and field annotations that can be attributed to a Michelson type corrspoding to t.

Constructors

NTc :: TypeAnn -> Notes (Tc ct)

NTKey :: TypeAnn -> Notes TKey

NTUnit :: TypeAnn -> Notes TUnit

NTSignature :: TypeAnn -> Notes TSignature

NTChainId :: TypeAnn -> Notes TChainId

NTOption :: TypeAnn -> Notes t -> Notes (TOption t)

NTList :: TypeAnn -> Notes t -> Notes (TList t)

NTSet :: TypeAnn -> TypeAnn -> Notes (TSet ct)

NTOperation :: TypeAnn -> Notes TOperation

NTContract :: TypeAnn -> Notes t -> Notes (TContract t)

NTPair :: TypeAnn -> FieldAnn -> FieldAnn -> Notes p -> Notes q -> Notes (TPair p q)

NTOr :: TypeAnn -> FieldAnn -> FieldAnn -> Notes p -> Notes q -> Notes (TOr p q)

NTLambda :: TypeAnn -> Notes p -> Notes q -> Notes (TLambda p q)

NTMap :: TypeAnn -> TypeAnn -> Notes v -> Notes (TMap k v)

NTBigMap :: TypeAnn -> TypeAnn -> Notes v -> Notes (TBigMap k v)

data AnnConvergeError where

Constructors

AnnConvergeError :: forall (tag :: Type). (Buildable (Annotation tag), Show (Annotation tag), Typeable tag) => Annotation tag -> Annotation tag -> AnnConvergeError

converge :: Notes t -> Notes t -> Either AnnConvergeError (Notes t)

Combines two annotations trees a and b into a new one c in such a way that c can be obtained from both a and b by replacing some empty leaves with type or/and field annotations.

convergeAnns :: forall (tag :: Type). (Buildable (Annotation tag), Show (Annotation tag), Typeable tag) => Annotation tag -> Annotation tag -> Either AnnConvergeError (Annotation tag)

Converge two type or field notes (which may be wildcards).

orAnn :: Annotation t -> Annotation t -> Annotation t

isStar :: SingI t => Notes t -> Bool

Checks if no annotations are present.

starNotes :: forall t. SingI t => Notes t

In memory of NStar constructor. Generates notes with no annotations.

notesSing :: SingI t => Notes t -> Sing t

Forget information about annotations, pick singleton with the same type.

Note: currently we cannot derive Sing from Notes without SingI because for comparable types notes do not remember which exact comparable was used.

notesT :: SingI t => Notes t -> T

Get term-level type of notes.

pattern NTcs :: SingI t => t ~ Tc ct => Sing ct -> Notes t

Similar to NTc pattern, but gives singleton of the comparable type.