Michelson.Typed.EntryPoints

Description

Utilities for lightweight entrypoints support.

Documentation

data EpAddress

Address with optional entrypoint name attached to it. TODO: come up with better name?

Constructors

EpAddress

Fields

eaAddress :: Address

Address itself

eaEntryPoint :: EpName

Entrypoint name (might be empty)

data ParseEpAddressError

Constructors

ParseEpAddressBadAddress ParseAddressError

ParseEpAddressBadRefAnn Text

ParseEpAddressRefAnnError EpNameFromRefAnnError

formatEpAddress :: EpAddress -> Text

mformatEpAddress :: EpAddress -> MText

parseEpAddress :: Text -> Either ParseEpAddressError EpAddress

Parse an address which can be suffixed with entrypoint name (e.g. "tz1faswCTDciRzE4oJ9jn2Vm2dvjeyA9fUzU%entrypoint").

unsafeParseEpAddress :: HasCallStack => Text -> EpAddress

newtype ParamNotes (t :: T)

Annotations for contract parameter declaration.

Following the Michelson specification, this type has the following invariants: 1. No entrypoint name is duplicated. 2. If default entrypoint is explicitly assigned, no "arm" remains uncallable.

Constructors

ParamNotesUnsafe

Fields

unParamNotes :: Notes t

pattern ParamNotes :: Notes t -> ParamNotes t

starParamNotes :: SingI t => ParamNotes t

Parameter without annotations.

data ArmCoord

Constructors

AcLeft

AcRight

type ArmCoords = [ArmCoord]

Coordinates of "arm" in Or tree, used solely in error messages.

data ParamEpError

Errors specific to parameter type declaration (entrypoints).

Constructors

ParamEpDuplicatedNames (NonEmpty EpName)

ParamEpUncallableArm ArmCoords

mkParamNotes :: Notes t -> Either ParamEpError (ParamNotes t)

Construct ParamNotes performing all necessary checks.

data EpLiftSequence (arg :: T) (param :: T) where

Describes how to construct full contract parameter from given entrypoint argument.

This could be just wrapper over Value arg -> Value param, but we cannot use Value type in this module easily.

Constructors

EplArgHere :: EpLiftSequence arg arg

EplWrapLeft :: EpLiftSequence arg subparam -> EpLiftSequence arg (TOr subparam r)

EplWrapRight :: EpLiftSequence arg subparam -> EpLiftSequence arg (TOr l subparam)

data EntryPointCallT (param :: T) (arg :: T)

Reference for calling a specific entrypoint of type arg.

Constructors

ParameterScope arg => EntryPointCall

Fields

epcName :: EpName

Name of entrypoint.

epcParamProxy :: Proxy param

Proxy of parameter, to make parameter type more easily fetchable.

epcLiftSequence :: EpLiftSequence arg param

How to call this entrypoint in the corresponding contract.

epcPrimitive :: forall p. (ParameterScope p, ForbidOr p) => EntryPointCallT p p

Call parameter which has no entrypoints, always safe.

epcCallRootUnsafe :: ParameterScope param => EntryPointCallT param param

Construct EntryPointCallT which calls no entrypoint and assumes that there is no explicit "default" one.

Validity of such operation is not ensured.

data SomeEntryPointCallT (arg :: T)

EntryPointCallT with hidden parameter type.

This requires argument to satisfy ParameterScope constraint. Strictly speaking, entrypoint argument may one day start having different set of constraints comparing to ones applied to parameter, but this seems unlikely.

Constructors

ParameterScope param => SomeEpc (EntryPointCallT param arg)

sepcCallRootUnsafe :: ParameterScope param => SomeEntryPointCallT param

Construct SomeEntryPointCallT which calls no entrypoint and assumes that there is no explicit "default" one.

Validity of such operation is not ensured.

sepcPrimitive :: forall t. (ParameterScope t, ForbidOr t) => SomeEntryPointCallT t

Call parameter which has no entrypoints, always safe.

sepcName :: SomeEntryPointCallT arg -> EpName

type family ForbidOr (t :: T) :: Constraint where ...

Equations

ForbidOr (TOr l r) = TypeError (Text "Cannot apply to sum type parameter " :<>: ShowType (TOr l r))

ForbidOr _ = ()

data MkEntryPointCallRes param where

Constructors

MkEntryPointCallRes :: ParameterScope arg => Notes arg -> EntryPointCallT param arg -> MkEntryPointCallRes param

mkEntryPointCall :: ParameterScope param => EpName -> ParamNotes param -> Maybe (MkEntryPointCallRes param)

Build EntryPointCallT.

Here we accept entrypoint name and type information for the parameter of target contract.

Returns Nothing if entrypoint is not found.

tyImplicitAccountParam :: ParamNotes TUnit

Parameter type of implicit account.

flattenEntryPoints :: SingI t => ParamNotes t -> Map EpName Type

Flatten a provided list of notes to a map of its entrypoints and its corresponding utype. Please refer to mkEntrypointsMap in regards to how duplicate entrypoints are handled.

Re-exports

newtype EpName

Entrypoint name.

Empty if this entrypoint is default one. Cannot be equal to "default", the reference implementation forbids that. Also, set of allowed characters should be the same as in annotations.

Constructors

EpNameUnsafe

Fields

unEpName :: Text

pattern DefEpName :: EpName

epNameFromParamAnn :: FieldAnn -> Maybe EpName

Make up EpName from annotation in parameter type declaration.

Returns Nothing if no entrypoint is assigned here.

epNameToParamAnn :: EpName -> FieldAnn

Turn entrypoint name into annotation for contract parameter declaration.

epNameFromRefAnn :: FieldAnn -> Either EpNameFromRefAnnError EpName

Make up EpName from annotation which is reference to an entrypoint (e.g. annotation in CONTRACT instruction).

Fails if annotation is invalid.

epNameToRefAnn :: EpName -> FieldAnn

Turn entrypoint name into annotation used as reference to entrypoint.

data EpNameFromRefAnnError

Constructors

InEpNameBadAnnotation FieldAnn