Michelson.Typed.Haskell.Value

Description

Conversions between haskell types/values and Michelson ones.

Value conversions

class IsoCValue a where

Isomorphism between Michelson primitive values and plain Haskell types.

Associated Types

type ToCT a :: CT

Type function that converts a regular Haskell type into a comparable type (which has kind CT).

Methods

toCVal :: a -> CValue (ToCT a)

Converts a single Haskell value into CVal representation.

fromCVal :: CValue (ToCT a) -> a

Converts a CVal value into a single Haskell value.

class IsoValue a where

Isomorphism between Michelson values and plain Haskell types.

Default implementation of this typeclass converts ADTs to Michelson "pair"s and "or"s.

Minimal complete definition

Nothing

Associated Types

type ToT a :: T

Type function that converts a regular Haskell type into a T type.

Methods

toVal :: a -> Value (ToT a)

Converts a Haskell structure into Value representation.

toVal :: (Generic a, GIsoValue (Rep a), ToT a ~ GValueType (Rep a)) => a -> Value (ToT a)

Converts a Haskell structure into Value representation.

fromVal :: Value (ToT a) -> a

Converts a Value into Haskell type.

fromVal :: (Generic a, GIsoValue (Rep a), ToT a ~ GValueType (Rep a)) => Value (ToT a) -> a

Converts a Value into Haskell type.

class GIsoValue (x :: Type -> Type)

Implements ADT conversion to Michelson value.

Thanks to Generic, Michelson representation will be a balanced tree; this reduces average access time in general case.

A drawback of such approach is that, in theory, in new GHC version generified representation may change; however, chances are small and I (martoon) believe that contract versions will change much faster anyway.

Minimal complete definition

gToValue, gFromValue

Associated Types

type GValueType x :: T

type family ToT' (t :: k) :: T where ...

Overloaded version of ToT to work on Haskell and T types.

Equations

ToT' (t :: T) = t

ToT' (t :: Type) = ToT t

data SomeIsoValue where

Hides some Haskell value put in line with Michelson Value.

Constructors

SomeIsoValue :: (Typeable a, IsoValue a) => a -> SomeIsoValue

newtype AnyIsoValue

Any Haskell value which can be converted to Michelson Value.

Constructors

AnyIsoValue (forall a. IsoValue a => a)

type IsComparable c = ToT c ~ Tc (ToCT c)

A useful property which holds for all CT types.

type GenericIsoValue t = (IsoValue t, Generic t, ToT t ~ GValueType (Rep t))

Whether Michelson representation of the type is derived via Generics.

Missing Haskell analogies to Michelson values

type EntryPointCall param arg = EntryPointCallT (ToT param) (ToT arg)

type SomeEntryPointCall arg = SomeEntryPointCallT (ToT arg)

data ContractRef (arg :: Type)

Since Contract name is used to designate contract code, lets call analogy of TContract type as follows.

Note that type argument always designates an argument of entrypoint. If a contract has explicit default entrypoint (and no root entrypoint), ContractRef referring to it can never have the entire parameter as its type argument.

Constructors

ContractRef

Fields

crAddress :: Address

crEntryPoint :: SomeEntryPointCall arg

coerceContractRef :: ToT a ~ ToT b => ContractRef a -> ContractRef b

Replace type argument of ContractAddr with isomorphic one.

contractRefToAddr :: ContractRef cp -> EpAddress

newtype BigMap k v

Constructors

BigMap

Fields

unBigMap :: Map k v

Stack conversion

type family ToTs (ts :: [Type]) :: [T] where ...

Type function to convert a Haskell stack type to T-based one.

Equations

ToTs '[] = '[]

ToTs (x ': xs) = ToT x ': ToTs xs

type family ToTs' (t :: [k]) :: [T] where ...

Overloaded version of ToTs to work on Haskell and T stacks.

Equations

ToTs' (t :: [T]) = t

ToTs' (a :: [Type]) = ToTs a

class IsoValuesStack (ts :: [Type]) where

Isomorphism between Michelson stack and its Haskell reflection.

Methods

toValStack :: Rec Identity ts -> Rec Value (ToTs ts)

fromValStack :: Rec Value (ToTs ts) -> Rec Identity ts

totsKnownLemma :: forall s. KnownList s :- KnownList (ToTs s)

totsAppendLemma :: forall a b. KnownList a => Dict (ToTs (a ++ b) ~ (ToTs a ++ ToTs b))