Michelson.Typed.Polymorphic

Description

Module, containing type classes for operating with Michelson values in the context of polymorphic stack type operations.

Documentation

class EDivOp (n :: CT) (m :: CT) where

Associated Types

type EDivOpRes n m :: CT

type EModOpRes n m :: CT

Methods

evalEDivOp :: CValue n -> CValue m -> Value' instr (TOption (TPair (Tc (EDivOpRes n m)) (Tc (EModOpRes n m))))

class MemOp (c :: T) where

Associated Types

type MemOpKey c :: CT

Methods

evalMem :: CValue (MemOpKey c) -> Value' cp c -> Bool

class MapOp (c :: T) where

Associated Types

type MapOpInp c :: T

type MapOpRes c :: T -> T

Methods

mapOpToList :: Value' instr c -> [Value' instr (MapOpInp c)]

mapOpFromList :: Value' instr c -> [Value' instr b] -> Value' instr (MapOpRes c b)

class IterOp (c :: T) where

Associated Types

type IterOpEl c :: T

Methods

iterOpDetachOne :: Value' instr c -> (Maybe (Value' instr (IterOpEl c)), Value' instr c)

class SizeOp (c :: T) where

Methods

evalSize :: Value' cp c -> Int

class GetOp (c :: T) where

Associated Types

type GetOpKey c :: CT

type GetOpVal c :: T

Methods

evalGet :: CValue (GetOpKey c) -> Value' cp c -> Maybe (Value' cp (GetOpVal c))

class UpdOp (c :: T) where

Associated Types

type UpdOpKey c :: CT

type UpdOpParams c :: T

Methods

evalUpd :: CValue (UpdOpKey c) -> Value' cp (UpdOpParams c) -> Value' cp c -> Value' cp c

class SliceOp (c :: T) where

Methods

evalSlice :: Natural -> Natural -> Value' cp c -> Maybe (Value' cp c)

class ConcatOp (c :: T) where

Methods

evalConcat :: Value' cp c -> Value' cp c -> Value' cp c

evalConcat' :: [Value' cp c] -> Value' cp c