Michelson.Typed.Print

Description

Types printing.

Moved in a separate module, because we print via converting to untyped T type to avoid duplication.

Documentation

buildStack :: [T] -> Builder

Format type stack in a pretty way.

Orphan instances

Buildable T

Instance details

Methods

build :: T -> Builder