Michelson.Typed.Scope

Description

Module, containing restrictions imposed by instruction or value scope.

Michelson have multiple restrictions on values, examples: * operation type cannot appear in parameter. * big_map type cannot appear in PUSH-able constants. * contract type cannot appear in type we UNPACK to.

Thus we declare multiple "scopes" - constraints applied in corresponding situations, for instance * ParameterScope; * StorageScope; * ConstantScope.

Also we separate multiple "classes" of scope-related constraints.

ParameterScope and similar ones are used within Michelson engine, they are understandable by GHC but produce not very clarifying errors.

ProperParameterBetterErrors and similar ones are middle-layer constraints, they produce human-readable errors but GHC cannot make conclusions from them. They are supposed to be used only by eDSLs to define their own high-level constraints.

Lorentz and other eDSLs may declare their own constraints, in most cases you should use them. For example see Constraints module.

Scopes

type ComparabilityScope t = (Typeable t, SingI t, Comparable t)

Alias for comparable types.

type ParameterScope t = (Typeable t, SingI t, HasNoOp t, HasNoNestedBigMaps t)

Alias for constraints which Michelson applies to parameter.

type StorageScope t = (Typeable t, SingI t, HasNoOp t, HasNoNestedBigMaps t, HasNoContract t)

Alias for constraints which Michelson applies to contract storage.

type ConstantScope t = (SingI t, HasNoOp t, HasNoBigMap t, HasNoContract t)

Alias for constraints which Michelson applies to pushed constants.

type PackedValScope t = (SingI t, HasNoOp t, HasNoBigMap t)

Alias for constraints which Michelson applies to packed values.

type UnpackedValScope t = (PackedValScope t, ConstantScope t)

Alias for constraints which Michelson applies to unpacked values.

It is different from PackedValScope, e.g. contract type cannot appear in a value we unpack to.

type PrintedValScope t = (SingI t, HasNoOp t)

Alias for constraints which are required for printing.

type ProperParameterBetterErrors t = (Typeable t, SingI t, ForbidOp t, ForbidNestedBigMaps t)

type ProperStorageBetterErrors t = (Typeable t, SingI t, ForbidOp t, ForbidNestedBigMaps t, ForbidContract t)

type ProperConstantBetterErrors t = (SingI t, ForbidOp t, ForbidBigMap t, ForbidContract t)

type ProperPackedValBetterErrors t = (SingI t, ForbidOp t, ForbidBigMap t)

type ProperUnpackedValBetterErrors t = (ProperPackedValBetterErrors t, ProperConstantBetterErrors t)

type ProperPrintedValBetterErrors t = (SingI t, ForbidOp t)

type ProperComparabilityBetterErrors t = (SingI t, ForbidNonComparable t, Comparable t)

properParameterEvi :: forall t. ProperParameterBetterErrors t :- ParameterScope t

properStorageEvi :: forall t. ProperStorageBetterErrors t :- StorageScope t

properConstantEvi :: forall t. ProperConstantBetterErrors t :- ConstantScope t

properPackedValEvi :: forall t. ProperPackedValBetterErrors t :- PackedValScope t

properUnpackedValEvi :: forall t. ProperUnpackedValBetterErrors t :- UnpackedValScope t

properPrintedValEvi :: forall t. ProperPrintedValBetterErrors t :- PrintedValScope t

newtype a :- b infixr 9

This is the type of entailment.

a :- b is read as a "entails" b.

With this we can actually build a category for Constraint resolution.

e.g.

Because Eq a is a superclass of Ord a, we can show that Ord a entails Eq a.

Because instance Ord a => Ord [a] exists, we can show that Ord a entails Ord [a] as well.

This relationship is captured in the :- entailment type here.

Since p :- p and entailment composes, :- forms the arrows of a Category of constraints. However, Category only became sufficiently general to support this instance in GHC 7.8, so prior to 7.8 this instance is unavailable.

But due to the coherence of instance resolution in Haskell, this Category has some very interesting properties. Notably, in the absence of IncoherentInstances, this category is "thin", which is to say that between any two objects (constraints) there is at most one distinguishable arrow.

This means that for instance, even though there are two ways to derive Ord a :- Eq [a], the answers from these two paths _must_ by construction be equal. This is a property that Haskell offers that is pretty much unique in the space of languages with things they call "type classes".

What are the two ways?

Well, we can go from Ord a :- Eq a via the superclass relationship, and then from Eq a :- Eq [a] via the instance, or we can go from Ord a :- Ord [a] via the instance then from Ord [a] :- Eq [a] through the superclass relationship and this diagram by definition must "commute".

Diagrammatically,

                   Ord a
               ins /     \ cls
                  v       v
            Ord [a]     Eq a
               cls \     / ins
                    v   v
                   Eq [a]

This safety net ensures that pretty much anything you can write with this library is sensible and can't break any assumptions on the behalf of library authors.

Constructors

Sub (a -> Dict b)

data BadTypeForScope

Constructors

BtNotComparable

BtIsOperation

BtHasBigMap

BtHasNestedBigMap

BtHasContract

class CheckScope (c :: Constraint) where

Should be present for common scopes.

Methods

checkScope :: Either BadTypeForScope (Dict c)

Check that constraint hold for a given type.

Implementation internals

class IsComparable t ~ True => Comparable t

Constraint which ensures that type is comparable.

class ContainsBigMap t ~ False => HasNoBigMap t

Constraint which ensures that bigmap does not appear in a given type.

class ContainsNestedBigMaps t ~ False => HasNoNestedBigMaps t

Constraint which ensures that there are no nested bigmaps.

class ContainsOp t ~ False => HasNoOp t

Constraint which ensures that operation type does not appear in a given type.

Not just a type alias in order to be able to partially apply it (e.g. in Each).

class ContainsContract t ~ False => HasNoContract t

Constraint which ensures that contract type does not appear in a given type.

type family ContainsBigMap (t :: T) :: Bool where ...

Whether this type contains TBigMap type.

Equations

ContainsBigMap (Tc _) = False

ContainsBigMap TKey = False

ContainsBigMap TUnit = False

ContainsBigMap TSignature = False

ContainsBigMap TChainId = False

ContainsBigMap (TOption t) = ContainsBigMap t

ContainsBigMap (TList t) = ContainsBigMap t

ContainsBigMap (TSet _) = False

ContainsBigMap TOperation = False

ContainsBigMap (TContract t) = ContainsBigMap t

ContainsBigMap (TPair a b) = ContainsBigMap a || ContainsBigMap b

ContainsBigMap (TOr a b) = ContainsBigMap a || ContainsBigMap b

ContainsBigMap (TLambda _ _) = False

ContainsBigMap (TMap _ v) = ContainsBigMap v

ContainsBigMap (TBigMap _ _) = True

type family ContainsNestedBigMaps (t :: T) :: Bool where ...

Whether this type contains a type with nested TBigMaps .

Nested big_maps (i.e. big_map which contains another big_map inside of it's value type). Are prohibited in all contexts. Some context such as PUSH, APPLY, PACK/UNPACK instructions are more strict because they doesn't work with big_map at all.

Equations

ContainsNestedBigMaps (Tc _) = False

ContainsNestedBigMaps TKey = False

ContainsNestedBigMaps TUnit = False

ContainsNestedBigMaps TSignature = False

ContainsNestedBigMaps TChainId = False

ContainsNestedBigMaps (TOption t) = ContainsNestedBigMaps t

ContainsNestedBigMaps (TList t) = ContainsNestedBigMaps t

ContainsNestedBigMaps (TSet _) = False

ContainsNestedBigMaps TOperation = False

ContainsNestedBigMaps (TContract t) = ContainsNestedBigMaps t

ContainsNestedBigMaps (TPair a b) = ContainsNestedBigMaps a || ContainsNestedBigMaps b

ContainsNestedBigMaps (TOr a b) = ContainsNestedBigMaps a || ContainsNestedBigMaps b

ContainsNestedBigMaps (TLambda _ _) = False

ContainsNestedBigMaps (TMap _ v) = ContainsNestedBigMaps v

ContainsNestedBigMaps (TBigMap _ v) = ContainsBigMap v

type ForbidOp t = FailOnOperationFound (ContainsOp t)

This is like HasNoOp, it raises a more human-readable error when t type is concrete, but GHC cannot make any conclusions from such constraint as it can for HasNoOp. Though, hopefully, it will someday: #11503.

Use this constraint in our eDSL.

type ForbidContract t = FailOnContractFound (ContainsContract t)

type ForbidBigMap t = FailOnBigMapFound (ContainsBigMap t)

type ForbidNestedBigMaps t = FailOnNestedBigMapsFound (ContainsNestedBigMaps t)

type family FailOnBigMapFound (enabled :: Bool) :: Constraint where ...

Report a human-readable error about TBigMap at a wrong place.

Equations

FailOnBigMapFound True = TypeError (Text "BigMaps are not allowed in this scope")

FailOnBigMapFound False = ()

type family FailOnNestedBigMapsFound (enabled :: Bool) :: Constraint where ...

Report a human-readable error that TBigMap contains another TBigMap

Equations

FailOnNestedBigMapsFound True = TypeError (Text "Nested BigMaps are not allowed")

FailOnNestedBigMapsFound False = ()

type family FailOnOperationFound (enabled :: Bool) :: Constraint where ...

Report a human-readable error about TOperation at a wrong place.

Equations

FailOnOperationFound True = TypeError (Text "Operations are not allowed in this scope")

FailOnOperationFound False = ()

type ForbidNonComparable t = FailOnNonComparable (IsComparable t)

data Comparability t where

Constructors

CanBeCompared :: Comparable t => Comparability t

CannotBeCompared :: IsComparable t ~ False => Comparability t

data OpPresence (t :: T)

Whether the type contains TOperation, with proof.

Constructors

ContainsOp t ~ True => OpPresent

ContainsOp t ~ False => OpAbsent

data ContractPresence (t :: T)

Constructors

ContainsContract t ~ True => ContractPresent

ContainsContract t ~ False => ContractAbsent

data BigMapPresence (t :: T)

Constructors

ContainsBigMap t ~ True => BigMapPresent

ContainsBigMap t ~ False => BigMapAbsent

data NestedBigMapsPresence (t :: T)

Constructors

ContainsNestedBigMaps t ~ True => NestedBigMapsPresent

ContainsNestedBigMaps t ~ False => NestedBigMapsAbsent

comparabilityPresence :: Sing t -> Maybe (Dict $ Comparable t)

checkComparability :: Sing t -> Comparability t

checkOpPresence :: Sing (ty :: T) -> OpPresence ty

Check at runtime whether the given type contains TOperation.

checkContractTypePresence :: Sing (ty :: T) -> ContractPresence ty

Check at runtime whether the given type contains TContract.

checkBigMapPresence :: Sing (ty :: T) -> BigMapPresence ty

Check at runtime whether the given type contains TBigMap.

checkNestedBigMapsPresence :: Sing (ty :: T) -> NestedBigMapsPresence ty

Check at runtime whether the given type contains TBigMap.

opAbsense :: Sing (t :: T) -> Maybe (Dict $ HasNoOp t)

Check at runtime that the given type does not contain TOperation.

contractTypeAbsense :: Sing (t :: T) -> Maybe (Dict $ HasNoContract t)

Check at runtime that the given type does not contain TContract.

bigMapAbsense :: Sing (t :: T) -> Maybe (Dict $ HasNoBigMap t)

Check at runtime that the given type does not containt TBigMap

nestedBigMapsAbsense :: Sing (t :: T) -> Maybe (Dict $ HasNoNestedBigMaps t)

Check at runtime that the given type does not contain nested TBigMap

forbiddenOp :: forall t a. (SingI t, ForbidOp t) => (HasNoOp t => a) -> a

Reify HasNoOp contraint from ForbidOp.

Left for backward compatibility.

forbiddenContractType :: forall t a. (SingI t, ForbidContract t) => (HasNoContract t => a) -> a

Reify HasNoContract contraint from ForbidContract.

forbiddenBigMap :: forall t a. (SingI t, ForbidBigMap t) => (HasNoBigMap t => a) -> a

forbiddenNestedBigMaps :: forall t a. (SingI t, ForbidNestedBigMaps t) => (HasNoNestedBigMaps t => a) -> a

Re-exports

withDict :: HasDict c e => e -> (c -> r) -> r

From a Dict, takes a value in an environment where the instance witnessed by the Dict is in scope, and evaluates it.

Essentially a deconstruction of a Dict into its continuation-style form.

Can also be used to deconstruct an entailment, a :- b, using a context a.

withDict :: Dict c -> (c => r) -> r
withDict :: a => (a :- c) -> (c => r) -> r

class SingI (a :: k) where

A SingI constraint is essentially an implicitly-passed singleton. If you need to satisfy this constraint with an explicit singleton, please see withSingI or the Sing pattern synonym.

Methods

sing :: Sing a

Produce the singleton explicitly. You will likely need the ScopedTypeVariables extension to use this method the way you want.