Michelson.Typed.Sing

Description

Module, providing singleton boilerplate for T and CT data types.

Some functions from Data.Singletons are provided alternative version here. Some instances which are usually generated with TH are manually implemented as they require some specific constraints, namely Typeable and/or Converge, not provided in instances generated by TH.

Documentation

data family Sing (a :: k) :: Type

The singleton kind-indexed data family.

withSomeSingT :: T -> (forall (a :: T). (Typeable a, SingI a) => Sing a -> r) -> r

Version of withSomeSing with Typeable constraint provided to processing function.

Required for not to erase these useful constraints when doing conversion from value of type T to its singleton representation.

withSomeSingCT :: CT -> (forall (a :: CT). (SingI a, Typeable a) => Sing a -> r) -> r

Version of withSomeSing with Typeable constraint provided to processing function.

Required for not to erase this useful constraint when doing conversion from value of type CT to its singleton representation.

fromSingT :: Sing (a :: T) -> T

Version of fromSing specialized for use with data instance Sing :: T -> Type which requires Typeable constraint for some of its constructors

fromSingCT :: Sing (a :: CT) -> CT

Orphan instances

SingKind CT

Instance details

Associated Types

type Demote CT = (r :: Type)

Methods

fromSing :: Sing a -> Demote CT

toSing :: Demote CT -> SomeSing CT

SingKind T

Instance details

Associated Types

type Demote T = (r :: Type)

Methods

fromSing :: Sing a -> Demote T

toSing :: Demote T -> SomeSing T

SingI CInt

Instance details

Methods

sing :: Sing CInt

SingI CNat

Instance details

Methods

sing :: Sing CNat

SingI CString

Instance details

Methods

sing :: Sing CString

SingI CBytes

Instance details

Methods

sing :: Sing CBytes

SingI CMutez

Instance details

Methods

sing :: Sing CMutez

SingI CBool

Instance details

Methods

sing :: Sing CBool

SingI CKeyHash

Instance details

Methods

sing :: Sing CKeyHash

SingI CTimestamp

Instance details

Methods

sing :: Sing CTimestamp

SingI CAddress

Instance details

Methods

sing :: Sing CAddress

SingI TKey

Instance details

Methods

sing :: Sing TKey

SingI TUnit

Instance details

Methods

sing :: Sing TUnit

SingI TSignature

Instance details

Methods

sing :: Sing TSignature

SingI TChainId

Instance details

Methods

sing :: Sing TChainId

SingI TOperation

Instance details

Methods

sing :: Sing TOperation

(SingI t, Typeable t) => SingI (Tc t :: T)

Instance details

Methods

sing :: Sing (Tc t)

(SingI a, Typeable a) => SingI (TOption a :: T)

Instance details

Methods

sing :: Sing (TOption a)

(SingI a, Typeable a) => SingI (TList a :: T)

Instance details

Methods

sing :: Sing (TList a)

(SingI a, Typeable a) => SingI (TSet a :: T)

Instance details

Methods

sing :: Sing (TSet a)

(SingI a, Typeable a) => SingI (TContract a :: T)

Instance details

Methods

sing :: Sing (TContract a)

(SingI a, Typeable a, Typeable b, SingI b) => SingI (TPair a b :: T)

Instance details

Methods

sing :: Sing (TPair a b)

(SingI a, Typeable a, Typeable b, SingI b) => SingI (TOr a b :: T)

Instance details

Methods

sing :: Sing (TOr a b)

(SingI a, Typeable a, Typeable b, SingI b) => SingI (TLambda a b :: T)

Instance details

Methods

sing :: Sing (TLambda a b)

(SingI a, Typeable a, Typeable b, SingI b) => SingI (TMap a b :: T)

Instance details

Methods

sing :: Sing (TMap a b)

(SingI a, Typeable a, Typeable b, SingI b) => SingI (TBigMap a b :: T)

Instance details

Methods

sing :: Sing (TBigMap a b)