Michelson.Typed.Util

Description

General-purpose utility functions for typed types.

Documentation

data DfsSettings x

Options for dfsInstr.

Constructors

DfsSettings

Fields

dsGoToValues :: Bool

Whether dfsInstr function should go into values which contain other instructions: lambdas and constant contracts (which can be passed to CREATE_CONTRACT).

dsCtorEffectsApp :: CtorEffectsApp x

How do we handle intermediate nodes in instruction tree.

data CtorEffectsApp x

Describes how intermediate nodes in instruction tree are accounted.

Constructors

CtorEffectsApp

Fields

ceaName :: Text

Name of this way.

ceaApplyEffects :: forall i o. Semigroup x => x -> x -> Instr i o -> (Instr i o, x)

This function accepts: 1. Effects gathered after applying step to node's children, but before applying it to the node itself. 2. Effects gathered after applying step to the given intermediate node. 3. Instruction resulting after all modifications produced by step.

dfsInstr :: forall x inp out. Semigroup x => DfsSettings x -> (forall i o. Instr i o -> (Instr i o, x)) -> Instr inp out -> (Instr inp out, x)

Traverse a typed instruction in depth-first order. <> is used to concatenate intermediate results. Each instructions can be changed using the supplied step function. It does not consider extra instructions (not present in Michelson).

dfsFoldInstr :: forall x inp out. Semigroup x => DfsSettings x -> (forall i o. Instr i o -> x) -> Instr inp out -> x

Specialization of dfsInstr for case when changing the instruction is not required.

dfsModifyInstr :: DfsSettings () -> (forall i o. Instr i o -> Instr i o) -> Instr inp out -> Instr inp out

Specialization of dfsInstr which only modifies given instruction.

Changing instruction tree structure

linearizeLeft :: Instr inp out -> Instr inp out

There are many ways to represent a sequence of more than 2 instructions. E. g. for `i1; i2; i3` it can be Seq i1 $ Seq i2 i3 or Seq (Seq i1 i2) i3. This function enforces a particular structure. Specifically, it makes each Instr have a single instruction (i. e. not Instr) in its second argument. This function also erases redundant Nops.

Value analysis

isStringValue :: Value t -> Maybe MText

If value is a string, return the stored string.

isBytesValue :: Value t -> Maybe ByteString

If value is a bytestring, return the stored bytestring.

allAtomicValues :: forall t a. (forall t'. Value t' -> Maybe a) -> Value t -> [a]

Takes a selector which checks whether an atomic value (i. e. that can not contain another value) can be converted to something. Recursively applies it to all atomic values in potentially non-atomic value. Collects extracted values in a list.

Perhaps one day we'll have dfsValue.