Michelson.Typed.Value

Description

Module, containing data types for Michelson value.

Documentation

data Value' instr t where

Representation of Michelson value.

Type parameter instr stands for Michelson instruction type, i.e. data type to represent an instruction of language.

Constructors

VC :: CValue t -> Value' instr (Tc t)

VKey :: PublicKey -> Value' instr TKey

VUnit :: Value' instr TUnit

VSignature :: Signature -> Value' instr TSignature

VChainId :: ChainId -> Value' instr TChainId

VOption :: forall t instr. Maybe (Value' instr t) -> Value' instr (TOption t)

VList :: forall t instr. [Value' instr t] -> Value' instr (TList t)

VSet :: forall t instr. Set (CValue t) -> Value' instr (TSet t)

VOp :: Operation' instr -> Value' instr TOperation

VContract :: forall arg instr. Address -> SomeEntryPointCallT arg -> Value' instr (TContract arg)

VPair :: forall l r instr. (Value' instr l, Value' instr r) -> Value' instr (TPair l r)

VOr :: forall l r instr. Either (Value' instr l) (Value' instr r) -> Value' instr (TOr l r)

VLam :: forall inp out instr. (forall i o. (Show (instr i o), Eq (instr i o), NFData (instr i o))) => RemFail instr (inp ': '[]) (out ': '[]) -> Value' instr (TLambda inp out)

VMap :: forall k v instr. Map (CValue k) (Value' instr v) -> Value' instr (TMap k v)

VBigMap :: forall k v instr. Map (CValue k) (Value' instr v) -> Value' instr (TBigMap k v)

data SomeValue' instr where

Constructors

SomeValue :: (Typeable t, SingI t) => Value' instr t -> SomeValue' instr

data SomeConstrainedValue' instr (c :: T -> Constraint) where

Constructors

SomeConstrainedValue :: forall (t :: T) (c :: T -> Constraint) instr. c t => Value' instr t -> SomeConstrainedValue' instr c

type ContractInp1 param st = TPair param st

type ContractInp param st = '[ContractInp1 param st]

type ContractOut1 st = TPair (TList TOperation) st

type ContractOut st = '[ContractOut1 st]

data CreateContract instr cp st

Constructors

(Show (instr (ContractInp cp st) (ContractOut st)), Eq (instr (ContractInp cp st) (ContractOut st))) => CreateContract

Fields

ccOriginator :: Address

ccDelegate :: Maybe KeyHash

ccBalance :: Mutez

ccStorageVal :: Value' instr st

ccContractCode :: instr (ContractInp cp st) (ContractOut st)

data CValue t where

Representation of comparable value in Michelson language.

By specification, we're allowed to compare only following types: int, nat, string, bytes, mutez, bool, key_hash, timestamp, address.

Only these values can be used as map keys or set elements.

Constructors

CvInt :: Integer -> CValue CInt

CvNat :: Natural -> CValue CNat

CvString :: MText -> CValue CString

CvBytes :: ByteString -> CValue CBytes

CvMutez :: Mutez -> CValue CMutez

CvBool :: Bool -> CValue CBool

CvKeyHash :: KeyHash -> CValue CKeyHash

CvTimestamp :: Timestamp -> CValue CTimestamp

CvAddress :: EpAddress -> CValue CAddress

data Operation' instr where

Data type, representing operation, list of which is returned by Michelson contract (according to calling convention).

These operations are to be further executed against system state after the contract execution.

Constructors

OpTransferTokens :: ParameterScope p => TransferTokens instr p -> Operation' instr

OpSetDelegate :: SetDelegate -> Operation' instr

OpCreateContract :: (Show (instr (ContractInp cp st) (ContractOut st)), NFData (instr (ContractInp cp st) (ContractOut st)), Typeable instr, ParameterScope cp, StorageScope st) => CreateContract instr cp st -> Operation' instr

data SetDelegate

Constructors

SetDelegate

Fields

sdMbKeyHash :: Maybe KeyHash

data TransferTokens instr p

Constructors

TransferTokens

Fields

ttTransferArgument :: Value' instr p

ttAmount :: Mutez

ttContract :: Value' instr (TContract p)

data RemFail (instr :: k -> k -> Type) (i :: k) (o :: k) where

Wrapper over instruction which remembers whether this instruction always fails or not.

Constructors

RfNormal :: instr i o -> RemFail instr i o

RfAlwaysFails :: (forall o'. instr i o') -> RemFail instr i o

rfMerge :: (forall o'. instr i1 o' -> instr i2 o' -> instr i3 o') -> RemFail instr i1 o -> RemFail instr i2 o -> RemFail instr i3 o

Merge two execution branches.

rfAnyInstr :: RemFail instr i o -> instr i o

Get code disregard whether it always fails or not.

rfMapAnyInstr :: (forall o'. instr i1 o' -> instr i2 o') -> RemFail instr i1 o -> RemFail instr i2 o

Modify inner code.

addressToVContract :: forall t instr. (ParameterScope t, ForbidOr t) => Address -> Value' instr (TContract t)

Make value of contract type which refers to the given address and does not call any entrypoint.

buildVContract :: Value' instr (TContract arg) -> Builder

compileEpLiftSequence :: EpLiftSequence arg param -> Value' instr arg -> Value' instr param

Turn EpLiftSequence into actual function on Values.

liftCallArg :: EntryPointCallT param arg -> Value' instr arg -> Value' instr param

Lift entrypoint argument to full parameter.