Michelson.Untyped.Annotation

Description

Michelson annotations in untyped model.

Documentation

newtype Annotation tag

Generic TypeFieldVariable Annotation

As per Michelson documentation, this type has an invariant: (except for the first character, here parametrized in the type tag) the allowed character set is the one matching the following regexp: %|%%|%|[:%][_a-zA-Z][_0-9a-zA-Z.%@]*

Constructors

AnnotationUnsafe

Fields

unAnnotation :: Text

pattern Annotation :: Text -> Annotation tag

pattern WithAnn :: Annotation tag -> Annotation tag

Annotation Set

data AnnotationSet

An AnnotationSet contains all the typefieldvariable Annotations , with each group in order, associated with an entity. Note that in its rendering/show instances the unnecessary annotations will be omitted, as well as in some of the functions operating with it. Necessary Annotations are the ones strictly required for a consistent representation. In particular, for each group (tfv): - if all annotations are noAnn they are all omitted - if one or more noAnn follow a non-empty ann, they are omitted - if one or more noAnn precede a non-empty ann, they are kept - every non-empty ann is obviously kept This is why order for each group is important as well as separation of different groups of Annotations.

emptyAnnSet :: AnnotationSet

An AnnotationSet without any Annotation.

fullAnnSet :: [TypeAnn] -> [FieldAnn] -> [VarAnn] -> AnnotationSet

An AnnotationSet built from all 3 kinds of Annotation.

isNoAnnSet :: AnnotationSet -> Bool

Returns True if all Annotations in the Set are unnecessaryemptynoAnn. False otherwise.

minAnnSetSize :: AnnotationSet -> Int

Returns the amount of Annotations that are necessary for a consistent representation. See AnnotationSet.

singleAnnSet :: forall tag. KnownAnnTag tag => Annotation tag -> AnnotationSet

An AnnotationSet with only a single Annotation (of any kind).

singleGroupAnnSet :: forall tag. KnownAnnTag tag => [Annotation tag] -> AnnotationSet

An AnnotationSet with several Annotations of the same kind.

Rendering

class Typeable (tag :: Type) => KnownAnnTag tag where

Methods

annPrefix :: Text

type TypeAnn = Annotation TypeTag

type FieldAnn = Annotation FieldTag

type VarAnn = Annotation VarTag

type SomeAnn = Annotation SomeTag

Creation and conversions

noAnn :: Annotation a

ann :: HasCallStack => Text -> Annotation a

Makes an Annotation from its textual value, prefix (%@:) excluded Throws an error if the given Text contains invalid characters

mkAnnotation :: Text -> Either Text (Annotation a)

Makes an Annotation from its textual value, prefix (%@:) excluded Returns a Text error message if the given Text contains invalid characters

specialVarAnns :: [Text]

List of all the special Variable Annotations, only allowed in CAR and CDR instructions, prefix (@) excluded. These do not respect the rules of isValidAnnStart and isValidAnnBodyChar.

specialFieldAnn :: Text

The only special Field Annotation, only allowed in PAIR, LEFT and RIGHT instructions, prefix (%) excluded. This does not respect the rules of isValidAnnStart and isValidAnnBodyChar.

isValidAnnStart :: Char -> Bool

Checks if a Char is valid to be the first of an annotation, prefix (%@:) excluded, the ones following should be checked with isValidAnnBodyChar instead. Note that this does not check Special Annotations, see specialVarAnns and specialFieldAnn

isValidAnnBodyChar :: Char -> Bool

Checks if a Char is valid to be part of an annotation, following a valid first character (see isValidAnnStart) and the prefix (%@:). Note that this does not check Special Annotations, see specialVarAnns and specialFieldAnn

unifyAnn :: Annotation tag -> Annotation tag -> Maybe (Annotation tag)

ifAnnUnified :: Annotation tag -> Annotation tag -> Bool

disjoinVn :: VarAnn -> (VarAnn, VarAnn)

convAnn :: Annotation tag1 -> Annotation tag2