Michelson.Untyped.EntryPoints

Documentation

newtype EpName

Entrypoint name.

Empty if this entrypoint is default one. Cannot be equal to "default", the reference implementation forbids that. Also, set of allowed characters should be the same as in annotations.

Constructors

EpNameUnsafe

Fields

unEpName :: Text

pattern DefEpName :: EpName

epNameFromParamAnn :: FieldAnn -> Maybe EpName

Make up EpName from annotation in parameter type declaration.

Returns Nothing if no entrypoint is assigned here.

epNameToParamAnn :: EpName -> FieldAnn

Turn entrypoint name into annotation for contract parameter declaration.

epNameFromRefAnn :: FieldAnn -> Either EpNameFromRefAnnError EpName

Make up EpName from annotation which is reference to an entrypoint (e.g. annotation in CONTRACT instruction).

Fails if annotation is invalid.

epNameToRefAnn :: EpName -> FieldAnn

Turn entrypoint name into annotation used as reference to entrypoint.

data EpNameFromRefAnnError

Constructors

InEpNameBadAnnotation FieldAnn

mkEntrypointsMap :: Type -> Map EpName Type

Given an untyped type, extract a map that maps entrypoint names to the their parameter types. If there are duplicate entrypoints in the given Type then the duplicate entrypoints at a deeper nesting level will get overwritten with the ones that are on top.