Michelson.Untyped.Ext

Documentation

data ExtInstrAbstract op

Implementation-specific instructions embedded in a NOP primitive, which mark a specific point during a contract's typechecking or execution.

These instructions are not allowed to modify the contract's stack, but may impose additional constraints that can cause a contract to report errors in type-checking or testing.

Additionaly, some implementation-specific language features such as type-checking of LetMacros are implemented using this mechanism (specifically FN and FN_END).

Constructors

STACKTYPE StackTypePattern
Matches current stack against a type-pattern

FN Text StackFn [op]
A typed stack function (push and pop a TcExtFrame)

UTEST_ASSERT (TestAssert op)
Copy the current stack and run an inline assertion on it

UPRINT PrintComment
Print a comment with optional embedded StackRefs

newtype StackRef

A reference into the stack.

Constructors

StackRef Natural

newtype PrintComment

Constructors

PrintComment

Fields

unUPrintComment :: [Either Text StackRef]

data TestAssert op

Constructors

TestAssert

Fields

tassName :: Text

tassComment :: PrintComment

tassInstrs :: [op]

newtype Var

Constructors

Var Text

data TyVar

A type-variable or a type-constant

Constructors

VarID Var

TyCon Type

data StackTypePattern

A stack pattern-match

Constructors

StkEmpty

StkRest

StkCons TyVar StackTypePattern

data StackFn

A stack function that expresses the type signature of a LetMacro

Constructors

StackFn

Fields

quantifiedVars :: Maybe (Set Var)

inPattern :: StackTypePattern

outPattern :: StackTypePattern

varSet :: StackTypePattern -> Set Var

Get the set of variables in a stack pattern

stackTypePatternToList :: StackTypePattern -> ([TyVar], Bool)

Convert StackTypePattern to a list of types. Also returns Bool which is True if the pattern is a fixed list of types and False if it's a pattern match on the head of the stack.