Michelson.Untyped.Instr

Description

Michelson instructions in untyped model.

Documentation

data InstrAbstract op

Michelson instruction with abstract parameter op. This parameter is necessary, because at different stages of our pipeline it will be different. Initially it can contain macros and non-flattened instructions, but then it contains only vanilla Michelson instructions.

Constructors

EXT (ExtInstrAbstract op)

DROPN Word
"DROP n" instruction. Note: reference implementation permits int16 here. Negative numbers are parsed successfully there, but rejected later. Morley is more permissive, so we use Word here, i. e. permit more positive numbers. We do not permit negative numbers at type level. In practice, probably nobody will ever have numbers greater than ≈1000 here, at least due to gas limits. Same reasoning applies to other instructions which have a numeric parameter representing number of elements on stack.

DROP
DROP is essentially as special case for DROPN, but we need both because they are packed differently.

DUP VarAnn

SWAP

DIG Word

DUG Word

PUSH VarAnn Type (Value' op)

SOME TypeAnn VarAnn

NONE TypeAnn VarAnn Type

UNIT TypeAnn VarAnn

IF_NONE [op] [op]

PAIR TypeAnn VarAnn FieldAnn FieldAnn

CAR VarAnn FieldAnn

CDR VarAnn FieldAnn

LEFT TypeAnn VarAnn FieldAnn FieldAnn Type

RIGHT TypeAnn VarAnn FieldAnn FieldAnn Type

IF_LEFT [op] [op]

NIL TypeAnn VarAnn Type

CONS VarAnn

IF_CONS [op] [op]

SIZE VarAnn

EMPTY_SET TypeAnn VarAnn Comparable

EMPTY_MAP TypeAnn VarAnn Comparable Type

EMPTY_BIG_MAP TypeAnn VarAnn Comparable Type

MAP VarAnn [op]

ITER [op]

MEM VarAnn

GET VarAnn

UPDATE VarAnn

IF [op] [op]

LOOP [op]

LOOP_LEFT [op]

LAMBDA VarAnn Type Type [op]

EXEC VarAnn

APPLY VarAnn

DIP [op]

DIPN Word [op]

FAILWITH

CAST VarAnn Type

RENAME VarAnn

PACK VarAnn

UNPACK TypeAnn VarAnn Type

CONCAT VarAnn

SLICE VarAnn

ISNAT VarAnn

ADD VarAnn

SUB VarAnn

MUL VarAnn

EDIV VarAnn

ABS VarAnn

NEG VarAnn

LSL VarAnn

LSR VarAnn

OR VarAnn

AND VarAnn

XOR VarAnn

NOT VarAnn

COMPARE VarAnn

EQ VarAnn

NEQ VarAnn

LT VarAnn

GT VarAnn

LE VarAnn

GE VarAnn

INT VarAnn

SELF VarAnn FieldAnn

CONTRACT VarAnn FieldAnn Type

TRANSFER_TOKENS VarAnn

SET_DELEGATE VarAnn

CREATE_CONTRACT VarAnn VarAnn (Contract' op)

IMPLICIT_ACCOUNT VarAnn

NOW VarAnn

AMOUNT VarAnn

BALANCE VarAnn

CHECK_SIGNATURE VarAnn

SHA256 VarAnn

SHA512 VarAnn

BLAKE2B VarAnn

HASH_KEY VarAnn

STEPS_TO_QUOTA VarAnn

SOURCE VarAnn

SENDER VarAnn

ADDRESS VarAnn

CHAIN_ID VarAnn

data ExpandedOp

Constructors

PrimEx ExpandedInstr

SeqEx [ExpandedOp]

WithSrcEx InstrCallStack ExpandedOp

type ExpandedInstr = InstrAbstract ExpandedOp

flattenExpandedOp :: ExpandedOp -> [ExpandedInstr]

Flatten all SeqEx in ExpandedOp. This function is mostly for testing. It returns instructions with the same logic, but they are not strictly equivalent, because they are serialized differently (grouping instructions into sequences affects the way they are PACK'ed).

Contract's address

data OriginationOperation

Data necessary to originate a contract.

Constructors

OriginationOperation

Fields

ooOriginator :: Address

Originator of the contract.

ooDelegate :: Maybe KeyHash

Optional delegate.

ooBalance :: Mutez

Initial balance of the contract.

ooStorage :: Value' ExpandedOp

Initial storage value of the contract.

ooContract :: Contract' ExpandedOp

The contract itself.

mkContractAddress :: OriginationOperation -> Address

Compute address of a contract from its origination operation.

TODO [TM-62] It's certainly imprecise, real Tezos implementation doesn't use JSON, but we don't need precise format yet, so we just use some serialization format (JSON because we have necessary instances already).