Michelson.Untyped.Type

Description

Michelson types represented in untyped model.

Documentation

data Type

Constructors

Type T TypeAnn

TypeParameter
Implicit Parameter type which can be used in contract code

TypeStorage
Implicit Storage type which can be used in contract code

data Comparable

Constructors

Comparable CT TypeAnn

compToType :: Comparable -> Type

typeToComp :: Type -> Maybe Comparable

data T

Constructors

Tc CT

TKey

TUnit

TSignature

TOption FieldAnn Type

TList Type

TSet Comparable

TOperation

TContract Type

TPair FieldAnn FieldAnn Type Type

TOr FieldAnn FieldAnn Type Type

TLambda Type Type

TMap Comparable Type

TBigMap Comparable Type

data CT

Constructors

CInt

CNat

CString

CBytes

CMutez

CBool

CKeyHash

CTimestamp

CAddress

pattern Tint :: T

pattern Tnat :: T

pattern Tstring :: T

pattern Tbytes :: T

pattern Tmutez :: T

pattern Tbool :: T

pattern Tkey_hash :: T

pattern Ttimestamp :: T

pattern Taddress :: T

tint :: T

tnat :: T

tstring :: T

tbytes :: T

tmutez :: T

tbool :: T

tkeyHash :: T

ttimestamp :: T

taddress :: T

toption :: Type -> T

tpair :: Type -> Type -> T

tor :: Type -> Type -> T

tyint :: Type

tynat :: Type

tyunit :: Type

tybool :: Type

typair :: Type -> Type -> Type

tyor :: Type -> Type -> Type

isAtomicType :: Type -> Bool

isKey :: Type -> Bool

isSignature :: Type -> Bool

isComparable :: Type -> Bool

isMutez :: Type -> Bool

isKeyHash :: Type -> Bool

isBool :: Type -> Bool

isString :: Type -> Bool

isInteger :: Type -> Bool

isTimestamp :: Type -> Bool

isNat :: Type -> Bool

isInt :: Type -> Bool

isBytes :: Type -> Bool