Morley.CLI

Description

Utilities for parsing Morley types using optparse-applicative.

Options

contractFileOption :: Parser FilePath

Parser for path to a contract code.

nowOption :: Parser (Maybe Timestamp)

Parser for the time returned by NOW instruction.

maxStepsOption :: Parser Word64

Parser for gas limit on contract execution.

dbPathOption :: Parser FilePath

Parser for path to database with Morley state.

txDataOption :: Parser TxData

Parser for transaction parameters.

keyHashOption :: Maybe KeyHash -> ("name" :! String) -> ("help" :! String) -> Parser KeyHash

Generic parser to read an option of KeyHash type.

valueOption :: Maybe Value -> ("name" :! String) -> ("help" :! String) -> Parser Value

Generic parser to read an option of Value type.

mutezOption :: Maybe Mutez -> ("name" :! String) -> ("help" :! String) -> Parser Mutez

Generic parser to read an option of Mutez type.

addressOption :: Maybe Address -> ("name" :! String) -> ("help" :! String) -> Parser Address

Generic parser to read an option of Address type.

onelineOption :: Parser Bool

--oneline flag.

entrypointOption :: ("name" :! String) -> ("help" :! String) -> Parser EpName

Generic parser to read an option of EpName type.

mTextOption :: Maybe MText -> ("name" :! String) -> ("help" :! String) -> Parser MText

Generic parser to read an option of MText type.

Orphan instances

HasCLReader Value

Instance details

Methods

getReader :: ReadM Value

getMetavar :: String