Morley.Micheline

Description

Module that provides type classes for converting to and from low-level Micheline representation.

Documentation

class ToExpression a where

Type class that provides an ability to convert something to Micheline Expression.

Methods

toExpression :: a -> Expression

class FromExpression a where

Type class that provides the ability to convert something from a Micheline Expression.

Methods

fromExpression :: Expression -> Maybe a