Tezos.Address

Description

Address in Tezos.

Documentation

newtype ContractHash

Hash of origination command for some contract.

Constructors

ContractHash ByteString

data Address

Data type corresponding to address structure in Tezos.

Constructors

KeyAddress KeyHash
tz address which is a hash of a public key.

ContractAddress ContractHash
KT address which corresponds to a callable contract.

mkKeyAddress :: PublicKey -> Address

Smart constructor for KeyAddress.

detGenKeyAddress :: ByteString -> Address

Deterministically generate a random KeyAddress and discard its secret key.

mkContractAddressRaw :: ByteString -> Address

Smart constructor for ContractAddress. Its argument is serialized origination operation.

Note: it's quite unsafe to pass ByteString, because we can pass some garbage which is not a serialized origination operation, but this operation includes contract itself and necessary types are defined in *. So we have to serialize this data outside this module and pass it here as a ByteString. Alternatively we could add some constraint, but it would be almost as unsafe as passing a ByteString. For this reason we add Raw suffix to this function and provide a safer function in Instr. We may reconsider it later.

mkContractHashRaw :: ByteString -> ContractHash

Create a dummy ContractHash value.

Formatting

data ParseAddressError

Errors that can happen during address parsing.

Constructors

ParseAddressWrongBase58Check
Address is not in Base58Check format.

ParseAddressBothFailed CryptoParseError ParseContractAddressError
Both address parsers failed with some error.

data ParseContractAddressError

formatAddress :: Address -> Text

mformatAddress :: Address -> MText

parseAddress :: Text -> Either ParseAddressError Address

Parse an address from its human-readable textual representation used by Tezos (e. g. "tz1faswCTDciRzE4oJ9jn2Vm2dvjeyA9fUzU"). Or fail if it's invalid.

unsafeParseAddress :: HasCallStack => Text -> Address

Partial version of parseAddress which assumes that the address is correct. Can be used in tests.

unsafeParseContractHash :: HasCallStack => Text -> ContractHash

Parse a TK contract address, fail if address does not match the expected format.