Tezos.Crypto.Hash

Description

Cryptographic primitives related to hashing.

Documentation

blake2b :: ByteString -> ByteString

Compute a cryptographic hash of a bytestring using the Blake2b_256 cryptographic hash function. It's used by the BLAKE2B instruction in Michelson.

blake2b160 :: ByteString -> ByteString

Compute a cryptographic hash of a bytestring using the Blake2b_160 cryptographic hash function.

sha256 :: ByteString -> ByteString

Compute a cryptographic hash of a bytestring using the Sha256 cryptographic hash function.

sha512 :: ByteString -> ByteString

Compute a cryptographic hash of a bytestring using the Sha512 cryptographic hash function.