Tezos.Crypto.Secp256k1

Description

Secp256k1 cryptographic primitives.

Cryptographic primitive types

data PublicKey

Secp256k1 public cryptographic key.

Constructors

PublicKey

Fields

unPublicKey :: PublicKey

pkBytes :: Maybe ByteString

This is the hack we use to make serialization correct. Decoding is currently not implemented, so when we have to decode bytes we remember these bytes and produce some random public key.

TODO (#18) remove it.

data SecretKey

Secp256k1 secret cryptographic key.

newtype Signature

Secp256k1 cryptographic signature.

Constructors

Signature

Fields

unSignature :: Signature

detSecretKey :: ByteString -> SecretKey

Deterministicaly generate a secret key from seed.

toPublic :: SecretKey -> PublicKey

Create a public key from a secret key.

Raw bytes (no checksums, tags or anything)

publicKeyToBytes :: forall ba. ByteArray ba => PublicKey -> ba

Convert a PublicKey to raw bytes.

TODO (#18): apparently it uses compressed SEC format as described in https://www.oreilly.com/library/view/programming-bitcoin/9781492031482/ch04.html However, it is not tested yet.

mkPublicKey :: ByteArrayAccess ba => ba -> Either CryptoParseError PublicKey

Make a PublicKey from raw bytes.

TODO (#18): it should decode from compressed SEC format, but it's left for a future task, so for now we return a constant.

publicKeyLengthBytes :: Integral n => n

signatureToBytes :: ByteArray ba => Signature -> ba

Convert a PublicKey to raw bytes.

TODO (#18): apparently a signature always has 64 bytes, so this format might be correct, but it is not tested.

mkSignature :: ByteArray ba => ba -> Either CryptoParseError Signature

Make a Signature from raw bytes.

TODO (#18): apparently a signature always has 64 bytes, so this format might be correct, but it is not tested.

signatureLengthBytes :: Integral n => n

Formatting and parsing

formatPublicKey :: PublicKey -> Text

mformatPublicKey :: PublicKey -> MText

parsePublicKey :: Text -> Either CryptoParseError PublicKey

formatSignature :: Signature -> Text

mformatSignature :: Signature -> MText

parseSignature :: Text -> Either CryptoParseError Signature

Signing

sign :: MonadRandom m => SecretKey -> ByteString -> m Signature

Sign a message using the secret key.

checkSignature :: PublicKey -> Signature -> ByteString -> Bool

Check that a sequence of bytes has been signed with a given key.