Tezos.Crypto.Util

Description

Utilities shared by multiple cryptographic primitives.

Documentation

data CryptoParseError

Error that can happen during parsing of cryptographic primitive types.

Constructors

CryptoParseWrongBase58Check

CryptoParseWrongTag ByteString

CryptoParseCryptoError CryptoError

CryptoParseUnexpectedLength Builder Int

encodeBase58Check :: ByteString -> Text

Encode a bytestring in Base58Check format.

decodeBase58Check :: Text -> Maybe ByteString

Decode a bytestring from Base58Check format.

data B58CheckWithPrefixError

Constructors

B58CheckWithPrefixWrongPrefix ByteString

B58CheckWithPrefixWrongEncoding

decodeBase58CheckWithPrefix :: ByteString -> Text -> Either B58CheckWithPrefixError ByteString

Parse a base58check encoded value expecting some prefix. If the actual prefix matches the expected one, it's stripped of and the resulting payload is returned.

formatImpl :: ByteArrayAccess x => ByteString -> x -> Text

Template for 'format*' functions.

parseImpl :: ByteString -> (ByteString -> Either CryptoParseError res) -> Text -> Either CryptoParseError res

Template for 'parse*' functions.

firstRight :: NonEmpty (Either e a) -> Either e a

Returns first encountered Right in a list. If there are none, returns arbitrary Left. It is useful to implement parsing.

deterministic :: ByteString -> MonadPseudoRandom ChaChaDRG a -> a

Do randomized action using specified seed.