Util.Exception

Common exceptions

data TextException

Constructors

TextException Text

Better printing of exceptions

displayUncaughtException :: IO () -> IO ()

Customise default uncaught exception handling. The problem with the default handler is that it uses show to display uncaught exceptions, but displayException may provide more reasonable output. We do not modify uncaught exception handler, but simply wrap uncaught exceptions (only synchronous ones) into DisplayExceptionInShow.

Some exceptions (currently we are aware only of ExitCode) are handled specially by default exception handler, so we don't wrap them.