Util.Generic

Description

Generic-related utils.

Documentation

mkGenericTree :: (Natural -> a -> a -> a) -> NonEmpty a -> a

Rebuild a list into a binary tree of exactly the same form which Generic uses to represent datatypes.

Along with the original list you have to provide constructor for intermediate nodes - it accepts zero-based index of the leftmost element of the right tree and merged trees themselves.

mkGenericTreeVec :: HasCallStack => (a -> b) -> (Natural -> b -> b -> b) -> Vector a -> b

type GenericTypeName a = GTypeName (Rep a)

Extract datatype name via its Generic representation.

For polymorphic types this throws away all type arguments.