Util.IO

Documentation

readFileUtf8 :: FilePath -> IO Text

writeFileUtf8 :: Print text => FilePath -> text -> IO ()

appendFileUtf8 :: Print text => FilePath -> text -> IO ()

withEncoding :: Handle -> TextEncoding -> IO () -> IO ()

hSetTranslit :: Handle -> IO ()

Change the character encoding of the given Handle to transliterate on unsupported characters instead of throwing an exception.