Util.Positive

Description

Definition of Positive type and related utilities.

Documentation

newtype Positive

Integer values starting from 1.

We define our own datatype in order to have Data instance for it, which can not be derived for third-party types without exported constructor.

Constructors

PositiveUnsafe

Fields

unPositive :: Natural

mkPositive :: (Integral i, Buildable i) => i -> Either Text Positive

lengthNE :: NonEmpty a -> Positive

Count length of non-empty list.

replicateNE :: Positive -> a -> NonEmpty a

Produce a non empty list consisting of the given value.