Util.Test.Ingredients

Description

Ingridients that we use in our test suite.

Documentation

ourIngredients :: [Ingredient]

This is the default set of ingredients extended with the antXMLRunner which is used to generate xml reports for CI.