Util.TypeLits

Description

Re-exports TypeLits, modifying it considering our practices.

Documentation

data Symbol

(Kind) This is the kind of type-level symbols. Declared here because class IP needs it

class KnownSymbol (n :: Symbol)

This class gives the string associated with a type-level symbol. There are instances of the class for every concrete literal: "hello", etc.

Since: base-4.7.0.0

Minimal complete definition

symbolSing

type family AppendSymbol (a :: Symbol) (b :: Symbol) :: Symbol where ...

Concatenation of type-level symbols.

Since: base-4.10.0.0

symbolVal :: KnownSymbol n => proxy n -> String

Since: base-4.7.0.0

symbolValT :: forall s. KnownSymbol s => Proxy s -> Text

symbolValT' :: forall s. KnownSymbol s => Text

type family TypeError (a :: ErrorMessage) :: b where ...

The type-level equivalent of error.

The polymorphic kind of this type allows it to be used in several settings. For instance, it can be used as a constraint, e.g. to provide a better error message for a non-existent instance,

-- in a context
instance TypeError (Text "Cannot Show functions." :$$:
                    Text "Perhaps there is a missing argument?")
      => Show (a -> b) where
    showsPrec = error "unreachable"

It can also be placed on the right-hand side of a type-level function to provide an error for an invalid case,

type family ByteSize x where
   ByteSize Word16   = 2
   ByteSize Word8    = 1
   ByteSize a        = TypeError (Text "The type " :<>: ShowType a :<>:
                                  Text " is not exportable.")

Since: base-4.9.0.0

data ErrorMessage where

A description of a custom type error.

Constructors

Text :: forall. Symbol -> ErrorMessage
Show the text as is.

ShowType :: forall t. t -> ErrorMessage
Pretty print the type. ShowType :: k -> ErrorMessage

(:<>:) :: forall. ErrorMessage -> ErrorMessage -> ErrorMessage infixl 6
Put two pieces of error message next to each other.

(:$$:) :: forall. ErrorMessage -> ErrorMessage -> ErrorMessage infixl 5
Stack two pieces of error message on top of each other.

type family TypeErrorUnless (cond :: Bool) (err :: ErrorMessage) :: Constraint where ...

Conditional type error.

Note that TypeErrorUnless cond err is the same as If cond () (TypeError err), but does not produce type-level error when one of its arguments cannot be deduced.

Equations

TypeErrorUnless True _ = ()

TypeErrorUnless False err = TypeError err

inTypeErrorUnless :: forall cond err a. TypeErrorUnless cond err => (cond ~ True => a) -> a

Reify the fact that condition under TypeErrorUnless constraint can be assumed to always hold.