Util.TypeTuple.Class

Documentation

class RecFromTuple r where

Building a record from tuple.

It differs from similar typeclass in FromTuple module in that it allows type inference outside-in - knowing desired Rec you know which tuple should be provided - this improves error messages when constructing concrete Rec objects.

Associated Types

type IsoRecTuple r :: Type

Methods

recFromTuple :: IsoRecTuple r -> r