Util.Typeable

Description

Utility for Typeable.

Documentation

gcastE :: forall a b t. (Typeable a, Typeable b) => t a -> Either Text (t b)

Like gcast, casts some container's elements, producing informative error on mismatch.

eqP :: (Typeable a, Typeable b) => Proxy a -> Proxy b -> Maybe (a :~: b)

Proxy version of eqT.

eqParam1 :: forall a1 a2 t. (Typeable a1, Typeable a2, Eq (t a1)) => t a1 -> t a2 -> Bool

Suppose you have a data type X with parameter a and you have two values: `x1 :: X a1` and `x2 :: X a2`. You can't compare them using ==, because they have different types. However, you can compare them using eqParam1 as long as both parameters are Typeable.

eqParam2 :: forall a1 a2 b1 b2 t. (Typeable a1, Typeable a2, Typeable b1, Typeable b2, Eq (t a1 b2)) => t a1 b1 -> t a2 b2 -> Bool

Version of eqParam1 for types with 2 parameters.

eqParam3 :: forall a1 a2 b1 b2 c1 c2 t. (Typeable a1, Typeable a2, Typeable b1, Typeable b2, Typeable c1, Typeable c2, Eq (t a1 b1 c1)) => t a1 b1 c1 -> t a2 b2 c2 -> Bool

Version of eqParam1 for types with 3 parameters.

eqExt :: forall a1 a2. (Typeable a1, Typeable a2, Eq a1) => a1 -> a2 -> Bool

Compare two entries of completely different types.

compareExt :: forall a1 a2. (Typeable a1, Typeable a2, Ord a1) => a1 -> a2 -> Ordering

Extension of eqExt to compare function.

castIgnoringPhantom :: forall c x. (Typeable x, Typeable c, forall phantom1 phantom2. Coercible (c phantom1) (c phantom2)) => x -> Maybe (c DummyPhantomType)

Cast to a type with phantom type argument without matching this argument. The phantom type must be the last type argument of the type.

Example of use: imagine a type

data MyType a = MyType

Normally, if object of this type was hidden under existential quantification with Typeable constraint, then in order to get it back with cast you need to know the exact type of the hidden object, including its phantom type parameter. With castIgnoringPhantom you get a way to extract this object no matter which phantom argument it had.

eqTypeIgnoringPhantom :: forall c x r. (Typeable x, Typeable c) => (forall a. Typeable a => (c a :~: x) -> Proxy a -> r) -> Maybe r

Match given type against another type of * -> * kind without caring about its type argument.

Re-exports

data (a :: k) :~: (b :: k) :: forall k. k -> k -> Type where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: base-4.7.0.0

Constructors

Refl :: forall k (a :: k) (b :: k). a :~: a

eqT :: (Typeable a, Typeable b) => Maybe (a :~: b)

Extract a witness of equality of two types

Since: base-4.7.0.0