Michelson.Interpret

Description

Module, containing function to interpret Michelson instructions against given context and input stack.

Documentation

data ContractEnv

Environment for contract execution.

Constructors

ContractEnv

Fields

ceNow :: !Timestamp

Timestamp returned by the NOW instruction.

ceMaxSteps :: !RemainingSteps

Number of steps after which execution unconditionally terminates.

ceBalance :: !Mutez

Current amount of mutez of the current contract.

ceContracts :: TcOriginatedContracts

Mapping from existing contracts' addresses to their executable representation.

ceSelf :: !Address

Address of the interpreted contract.

ceSource :: !Address

The contract that initiated the current transaction.

ceSender :: !Address

The contract that initiated the current internal transaction.

ceAmount :: !Mutez

Amount of the current transaction.

data InterpreterState

Constructors

InterpreterState

Fields

isMorleyLogs :: MorleyLogs

isRemainingSteps :: RemainingSteps

data MichelsonFailed where

Represents `[FAILED]` state of a Michelson program. Contains value that was on top of the stack when FAILWITH was called.

Constructors

MichelsonFailedWith :: (Typeable t, SingI t) => Value t -> MichelsonFailed

MichelsonArithError :: (Typeable n, Typeable m) => ArithError (CValue n) (CValue m) -> MichelsonFailed

MichelsonGasExhaustion :: MichelsonFailed

MichelsonFailedTestAssert :: Text -> MichelsonFailed

newtype RemainingSteps

Constructors

RemainingSteps Word64

data SomeItStack where

Constructors

SomeItStack :: ExtInstr inp -> Rec Value inp -> SomeItStack

type EvalOp a = ExceptT MichelsonFailed (ReaderT ContractEnv (State InterpreterState)) a

newtype MorleyLogs

Morley interpreter state

Constructors

MorleyLogs

Fields

unMorleyLogs :: [Text]

noMorleyLogs :: MorleyLogs

interpret :: Contract cp st -> Value cp -> Value st -> ContractEnv -> ContractReturn st

interpretRepeated :: Contract cp st -> [Value cp] -> Value st -> ContractEnv -> ContractReturn st

Emulate multiple calls of a contract.

interpretInstr :: ContractEnv -> Instr inp out -> Rec Value inp -> Either MichelsonFailed (Rec Value out)

Interpret an instruction in vacuum, putting no extra contraints on its execution.

Mostly for testing purposes.

type ContractReturn st = (Either MichelsonFailed ([Operation], Value st), InterpreterState)

interpretSome :: SomeContract -> SomeValue -> SomeValue -> ContractEnv -> Either InterpretError InterpretResult

Interpret a contract without performing any side effects using typed representation of contract, parameter and storage.

interpretUntyped :: Contract -> Value -> Value -> ContractEnv -> Either InterpretError InterpretResult

Interpret a contract without performing any side effects. This function uses untyped representation of contract, parameter and storage. Mostly used for testing.

data InterpretError

Constructors

RuntimeFailure (MichelsonFailed, MorleyLogs)

IllTypedContract TCError

IllTypedParam TCError

IllTypedStorage TCError

UnexpectedParamType TCTypeError

UnexpectedStorageType TCTypeError

data InterpretResult where

Constructors

InterpretResult

Fields

:: (Typeable st, SingI st, HasNoOp st)

=> { iurOps :: [Operation]

, iurNewStorage :: Value st

, iurNewState :: InterpreterState

} -> InterpretResult

runInstr :: Instr inp out -> Rec Value inp -> EvalOp (Rec Value out)

Function to change amount of remaining steps stored in State monad

runInstrNoGas :: forall a b. Instr a b -> Rec Value a -> EvalOp (Rec Value b)

runUnpack :: forall t. (SingI t, HasNoOp t, HasNoBigMap t) => TcOriginatedContracts -> ByteString -> Either UnpackError (Value t)

Unpacks given raw data into a typed value.