Michelson.Optimizer

Description

Optimizer for typed instructions.

It's quite experimental and incomplete. List of possible improvements: 1. See comment to optimizeWithRule, it is not quite correct. 2. pushDrop, dupDrop, unitDrop rules are essentially the same. It would be good to generalize them into one rule. The same applies to pushDip. 3. Also we completely ignore Nested. 4. Some rules make break right-linear structure (e. g. pushDip) and currently we use linearizeRight to account for that. It probably can be done more efficiently.

Documentation

optimize :: Instr inp out -> Instr inp out

Optimize a typed instruction by replacing some sequences of instructions with smaller equivalent sequences. Applies default set of rewrite rules.

optimizeWithRules :: (forall inp1 out1. Rules inp1 out1) -> Instr inp out -> Instr inp out

Optimize a typed instruction using a custom set of rules.

optimizeWithRule :: Rule -> Instr a b -> Instr a b

Optimize a typed instruction using a specific rule.

It applies all rules to head (first few instructions) while they successfully apply. Then it applies them to instructions after the first one and so on.

It's probably not the most efficient solution and it's not quite correct, because if we have instructions `a; b; c` and optimize `b; c` then we should return to optimizing sequence which starts with a.

linearizeRight :: Instr a b -> Instr a b

There are many ways to represent a sequence of more than 2 instructions. E. g. for `i1; i2; i3` it can be Seq i1 $ Seq i2 i3 or Seq i1 i2 $ i3. This function enforces a particular structure. Specifically, it makes each Instr have a single instruction (i. e. not Instr) in its first argument. It makes it easier to implement optimizer. This function also erases redundant Nop.

TODO: currently it doesn't go deeper into instructions like IF. We should try reusing deepRule for this purpose.

TODO: it also has bad worst-case performance, it should be carefully rewritten.

defaultRules :: Rules inp out