Michelson.TypeCheck.Error

Documentation

data TCTypeError

Data type that represents various errors which are related to type system. These errors are used to specify info about type check errors in TCError data type.

Constructors

AnnError AnnConvergeError
Annotation unify error

ExtractionTypeMismatch TypeConvergeError
Notes extraction error

TypeEqError T T
Type equality error

StackEqError [T] [T]
Stacks equality error

UnsupportedTypes [T]
Error that happens when some instruction doesn't have support for some types

UnknownType T
Error that happens when we meet unknown type

data TCError

Type check error

Constructors

TCFailedOnInstr ExpandedInstr SomeHST Text InstrCallStack (Maybe TCTypeError)

TCFailedOnValue Value T Text InstrCallStack (Maybe TCTypeError)

TCContractError Text (Maybe TCTypeError)

TCUnreachableCode InstrCallStack (NonEmpty ExpandedOp)

TCExtError SomeHST InstrCallStack ExtError

data ExtError

Various type errors possible when checking Morley extension commands

Constructors

LengthMismatch StackTypePattern

VarError Text StackFn

TypeMismatch StackTypePattern Int TCTypeError

TyVarMismatch Var Type StackTypePattern Int TCTypeError

StkRestMismatch StackTypePattern SomeHST SomeHST TCTypeError

TestAssertError Text

InvalidStackReference StackRef StackSize

newtype StackSize

Constructors

StackSize Natural