Michelson.Typed.Arith

Description

Module, containing some boilerplate for support of arithmetic operations in Michelson language.

Documentation

class ArithOp aop (n :: CT) (m :: CT) where

Class for binary arithmetic operation.

Takes binary operation marker as op parameter, types of left operand n and right operand m.

Associated Types

type ArithRes aop n m :: CT

Type family ArithRes denotes the type resulting from computing operation op from operands of types n and m.

For instance, adding integer to natural produces integer, which is reflected in following instance of type family: ArithRes Add CNat CInt = CInt.

Methods

evalOp :: proxy aop -> CValue n -> CValue m -> Either (ArithError (CValue n) (CValue m)) (CValue (ArithRes aop n m))

Evaluate arithmetic operation on given operands.

class UnaryArithOp aop (n :: CT) where

Marker data type for add operation.

Associated Types

type UnaryArithRes aop n :: CT

Methods

evalUnaryArithOp :: proxy aop -> CValue n -> CValue (UnaryArithRes aop n)

data ArithError n m

Represents an arithmetic error of the operation.

Constructors

MutezArithError ArithErrorType n m

ShiftArithError ArithErrorType n m

data ArithErrorType

Denotes the error type occured in the arithmetic operation.

Constructors

AddOverflow

MulOverflow

SubUnderflow

LslOverflow

LsrUnderflow

class CompareOp n where

Class for comparison operations, special case of ArithOp.

Methods

compareOp :: CValue n -> CValue n -> Integer

Evaluate compare operation.

data Add

data Sub

data Mul

data Abs

data Neg

data Or

data And

data Xor

data Not

data Lsl

data Lsr

data Compare

data Eq'

data Neq

data Lt

data Gt

data Le

data Ge