Michelson.Typed.Doc

Description

Renderable documentation injected to contract code.

Documentation

class (Typeable d, KnownNat (DocItemPosition d)) => DocItem d where

A piece of documentation describing one property of a thing, be it a name or description of a contract, or an error throwable by given endpoint.

Items of the same type appear close to each other in a rendered documentation and form a section.

Doc items are later injected into a contract code via a dedicated nop-like instruction. Normally doc items which belong to one section appear in resulting doc in the same order in which they appeared in the contract.

Minimal complete definition

docItemSectionName, docItemToMarkdown

Associated Types

type DocItemPosition d = (pos :: Nat) | pos -> d

Position of this item in the resulting documentation; the smaller the value, the higher the section with this element will be placed.

Documentation structure is not necessarily flat. If some doc item consolidates a whole documentation block within it, this block will have its own placement of items independent from outer parts of the doc.

type DocItemPlacement d :: DocItemPlacementKind

Defines where given doc item should be put. There are two options: 1. Inline right here (default behaviour); 2. Put into definitions section.

Methods

docItemSectionName :: Maybe Text

When multiple items of the same type belong to one section, how this section will be called.

If not provided, section will contain just untitled content.

docItemSectionDescription :: Maybe Builder

Description of a section.

Can be used to mention some common things about all elements of this section. Markdown syntax is permitted here.

docItemRef :: d -> DocItemRef (DocItemPlacement d)

Defines a function which constructs an unique identifier of given doc item, if it has been decided to put the doc item into definitions section.

Identifier should be unique both among doc items of the same type and items of other types. Thus, consider using "typeId-contentId" pattern.

docItemRef :: DocItemPlacement d ~ DocItemInlined => d -> DocItemRef (DocItemPlacement d)

Defines a function which constructs an unique identifier of given doc item, if it has been decided to put the doc item into definitions section.

Identifier should be unique both among doc items of the same type and items of other types. Thus, consider using "typeId-contentId" pattern.

docItemToMarkdown :: HeaderLevel -> d -> Builder

Render given doc item to Markdown, preferably one line, optionally with header.

Accepts the smallest allowed level of header. (Using smaller value than provided one will interfere with existing headers thus delivering mess).

docItemDependencies :: d -> [SomeDocDefinitionItem]

All doc items which this doc item refers to.

They will automatically be put to definitions as soon as given doc item is detected.

docItemsOrder :: [d] -> [d]

This function accepts doc items put under the same section in the order in which they appeared in the contract and returns their new desired order. It's also fine to use this function for filtering or merging doc items.

Default implementation * leaves inlined items as is; * for items put to definitions, lexicographically sorts them by their id.

docItemPosition :: forall d. DocItem d => DocItemPos

Get doc item position at term-level.

newtype DocItemId

Some unique identifier of a doc item.

All doc items which should be refer-able need to have this identifier.

Constructors

DocItemId Text

data DocItemPlacementKind

Where do we place given doc item.

Constructors

DocItemInlined
Placed in the document content itself.

DocItemInDefinitions
Placed in dedicated definitions section; can later be referenced.

data DocItemRef (p :: DocItemPlacementKind) where

Defines an identifier which given doc item can be referenced with.

Constructors

DocItemRef :: DocItemId -> DocItemRef DocItemInDefinitions

DocItemNoRef :: DocItemRef DocItemInlined

data SomeDocItem where

Hides some documentation item.

Constructors

SomeDocItem :: DocItem d => d -> SomeDocItem

data SomeDocDefinitionItem where

Hides some documentation item which is put to "definitions" section.

Constructors

SomeDocDefinitionItem :: (DocItem d, DocItemPlacement d ~ DocItemInDefinitions) => d -> SomeDocDefinitionItem

type DocBlock = Map DocItemPos (NonEmpty SomeDocItem)

A map from positions to document elements.

Note that each value in this map keeps a list of doc items, all of which have the same type (since each doc item type is forced to have unique position).

type DocGrouping = DocBlock -> SomeDocItem

A function which groups a piece of doc under one doc item.

data ContractDoc

Keeps documentation gathered for some piece of contract code.

Used for building documentation of a contract.

Constructors

ContractDoc

Fields

cdContents :: DocBlock

All inlined doc items.

cdDefinitions :: DocBlock

Definitions used in document.

Usually you put some large and repetitive descriptions here. This differs from the document content in that it contains sections which are always at top-level, disregard the nesting.

All doc items which define docItemId method go here, and only they.

cdDefinitionIds :: Set DocItemId

We remember all already used identifiers to avoid repetitions.

cdContentsL :: Lens' ContractDoc DocBlock

cdDefinitionsL :: Lens' ContractDoc DocBlock

cdDefinitionIdsL :: Lens' ContractDoc (Set DocItemId)

docBlockToMarkdown :: HeaderLevel -> DocBlock -> Builder

Render a documentation block.

contractDocToMarkdown :: ContractDoc -> LText

Render given contract documentation to markdown document.

data DName

Give a name to document block.

Constructors

DName Text DocBlock

data DDescription

Description of something.

Constructors

DDescription Text

data DVersion

Specify version if given contract.

Constructors

DVersion Natural

data DError

An error thrown by a contract.

Constructors

DError