Michelson.Typed.Util

Description

General-purpose utility functions for typed types.

Documentation

dfsInstr :: forall x inp out. Semigroup x => Bool -> (forall i o. Instr i o -> (Instr i o, x)) -> Instr inp out -> (Instr inp out, x)

Traverse a typed instruction in depth-first order. <> is used to concatenate intermediate results. First argument specifies whether this function should go into values which contains other instructions: lambdas and constant contracts (which can be passed to CREATE_CONTRACT). Each instructions can be changed using the supplied function. It does not consider extra instructions (not present in Michelson).