Michelson.Untyped.Annotation

Description

Michelson annotations in untyped model.

Documentation

newtype Annotation tag

Constructors

Annotation Text

pattern WithAnn :: Annotation tag -> Annotation tag

type TypeAnn = Annotation TypeTag

type FieldAnn = Annotation FieldTag

type VarAnn = Annotation VarTag

class RenderAnn t where

Typeclass for printing annotations, renderAnn prints empty prefix in case of noAnn.

Such functionality is required in case when instruction has two annotations of the same type, former is empty and the latter is not. So that `PAIR noAnn noAnn noAnn %kek` is printed as `PAIR % %kek`

Methods

renderAnn :: t -> Doc

noAnn :: Annotation a

ann :: Text -> Annotation a

unifyAnn :: Annotation tag -> Annotation tag -> Maybe (Annotation tag)

ifAnnUnified :: Annotation tag -> Annotation tag -> Bool

disjoinVn :: VarAnn -> (VarAnn, VarAnn)

convAnn :: Annotation tag1 -> Annotation tag2