Michelson.Untyped.Instr

Description

Michelson instructions in untyped model.

Documentation

data InstrAbstract op

Michelson instruction with abstract parameter op. This parameter is necessary, because at different stages of our pipeline it will be different. Initially it can contain macros and non-flattened instructions, but then it contains only vanilla Michelson instructions.

Constructors

EXT (ExtInstrAbstract op)

DROP

DUP VarAnn

SWAP

PUSH VarAnn Type (Value' op)

SOME TypeAnn VarAnn FieldAnn

NONE TypeAnn VarAnn FieldAnn Type

UNIT TypeAnn VarAnn

IF_NONE [op] [op]

PAIR TypeAnn VarAnn FieldAnn FieldAnn

CAR VarAnn FieldAnn

CDR VarAnn FieldAnn

LEFT TypeAnn VarAnn FieldAnn FieldAnn Type

RIGHT TypeAnn VarAnn FieldAnn FieldAnn Type

IF_LEFT [op] [op]

NIL TypeAnn VarAnn Type

CONS VarAnn

IF_CONS [op] [op]

SIZE VarAnn

EMPTY_SET TypeAnn VarAnn Comparable

EMPTY_MAP TypeAnn VarAnn Comparable Type

MAP VarAnn [op]

ITER [op]

MEM VarAnn

GET VarAnn

UPDATE

IF [op] [op]

LOOP [op]

LOOP_LEFT [op]

LAMBDA VarAnn Type Type [op]

EXEC VarAnn

DIP [op]

FAILWITH

CAST VarAnn Type

RENAME VarAnn

PACK VarAnn

UNPACK VarAnn Type

CONCAT VarAnn

SLICE VarAnn

ISNAT VarAnn

ADD VarAnn

SUB VarAnn

MUL VarAnn

EDIV VarAnn

ABS VarAnn

NEG

LSL VarAnn

LSR VarAnn

OR VarAnn

AND VarAnn

XOR VarAnn

NOT VarAnn

COMPARE VarAnn

EQ VarAnn

NEQ VarAnn

LT VarAnn

GT VarAnn

LE VarAnn

GE VarAnn

INT VarAnn

SELF VarAnn

CONTRACT VarAnn Type

TRANSFER_TOKENS VarAnn

SET_DELEGATE VarAnn

CREATE_ACCOUNT VarAnn VarAnn

CREATE_CONTRACT VarAnn VarAnn (Contract' op)

IMPLICIT_ACCOUNT VarAnn

NOW VarAnn

AMOUNT VarAnn

BALANCE VarAnn

CHECK_SIGNATURE VarAnn

SHA256 VarAnn

SHA512 VarAnn

BLAKE2B VarAnn

HASH_KEY VarAnn

STEPS_TO_QUOTA VarAnn

SOURCE VarAnn

SENDER VarAnn

ADDRESS VarAnn

newtype Op

Constructors

Op

Fields

unOp :: Instr

type Instr = InstrAbstract Op

data ExpandedOp

Constructors

PrimEx ExpandedInstr

SeqEx [ExpandedOp]

WithSrcEx InstrCallStack ExpandedOp

type ExpandedInstr = InstrAbstract ExpandedOp

type InstrExtU = ExtInstrAbstract Op

type ExpandedInstrExtU = ExtInstrAbstract ExpandedOp

flattenExpandedOp :: ExpandedOp -> [ExpandedInstr]

Flatten all SeqEx in ExpandedOp. This function is mostly for testing. It returns instructions with the same logic, but they are not strictly equivalent, because they are serialized differently (grouping instructions into sequences affects the way they are PACK'ed).

Note: it does not return a list of Instr because this type is not used anywhere and should probably be removed.

Contract's address

data OriginationOperation

Data necessary to originate a contract.

Constructors

OriginationOperation

Fields

ooManager :: !KeyHash

Manager of the contract.

ooDelegate :: !(Maybe KeyHash)

Optional delegate.

ooSpendable :: !Bool

Whether the contract is spendable.

ooDelegatable :: !Bool

Whether the contract is delegatable.

ooBalance :: !Mutez

Initial balance of the contract.

ooStorage :: !(Value' ExpandedOp)

Initial storage value of the contract.

ooContract :: !(Contract' ExpandedOp)

The contract itself.

mkContractAddress :: OriginationOperation -> Address

Compute address of a contract from its origination operation.

TODO [TM-62] It's certainly imprecise, real Tezos implementation doesn't use JSON, but we don't need precise format yet, so we just use some serialization format (JSON because we have necessary instances already).