Tezos.Address

Description

Address in Tezos.

Documentation

data Address

Data type corresponding to address structure in Tezos.

Constructors

KeyAddress !KeyHash
tz address which is a hash of a public key.

ContractAddress !ByteString
KT address which corresponds to a callable contract. It's a hash of origination command. TODO: we should probably have a Hash type.

mkKeyAddress :: PublicKey -> Address

Smart constructor for KeyAddress.

mkContractAddressRaw :: ByteString -> Address

Smart constructor for ContractAddress. Its argument is serialized origination operation.

Note: it's quite unsafe to pass ByteString, because we can pass some garbage which is not a serialized origination operation, but this operation includes contract itself and necessary types are defined in *. So we have to serialize this data outside this module and pass it here as a ByteString. Alternatively we could add some constraint, but it would be almost as unsafe as passing a ByteString. For this reason we add Raw suffix to this function and provide a safer function in Instr. We may reconsider it later.

Formatting

formatAddress :: Address -> Text

mformatAddress :: Address -> MText

parseAddress :: Text -> Either ParseAddressError Address

Parse an address from its human-readable textual representation used by Tezos (e. g. "tz1faswCTDciRzE4oJ9jn2Vm2dvjeyA9fUzU"). Or fail if it's invalid.

unsafeParseAddress :: HasCallStack => Text -> Address

Partial version of parseAddress which assumes that the address is correct. Can be used in tests.