Tezos.Core

Description

Core primitive Tezos types.

Mutez

data Mutez

Mutez is a wrapper over integer data type. 1 mutez is 1 token (μTz).

mkMutez :: Word64 -> Maybe Mutez

Safely create Mutez checking for overflow.

unsafeMkMutez :: HasCallStack => Word64 -> Mutez

Partial function for Mutez creation, it's pre-condition is that the argument must not exceed the maximal Mutez value.

toMutez :: Word32 -> Mutez

Safely create Mutez.

This is recommended way to create Mutez from a numeric literal; you can't construct all valid Mutez values using this function but for small values it works neat.

Warnings displayed when trying to construct invalid Natural or Word literal are hardcoded for these types in GHC implementation, so we can only exploit these existing rules.

addMutez :: Mutez -> Mutez -> Maybe Mutez

Addition of Mutez values. Returns Nothing in case of overflow.

unsafeAddMutez :: HasCallStack => Mutez -> Mutez -> Mutez

Partial addition of Mutez, should be used only if you're sure there'll be no overflow.

subMutez :: Mutez -> Mutez -> Maybe Mutez

Subtraction of Mutez values. Returns Nothing when the subtrahend is greater than the minuend, and Just otherwise.

unsafeSubMutez :: HasCallStack => Mutez -> Mutez -> Mutez

Partial subtraction of Mutez, should be used only if you're sure there'll be no underflow.

mulMutez :: Integral a => Mutez -> a -> Maybe Mutez

Multiplication of Mutez and an integral number. Returns Nothing in case of overflow.

divModMutez :: Mutez -> Mutez -> Maybe (Word64, Mutez)

Euclidian division of two Mutez values.

divModMutezInt :: Integral a => Mutez -> a -> Maybe (Mutez, Mutez)

Euclidian division of Mutez and a number.

Timestamp

newtype Timestamp

Time in the real world. Use the functions below to convert it to/from Unix time in seconds.

Constructors

Timestamp

Fields

unTimestamp :: POSIXTime

timestampToSeconds :: Integral a => Timestamp -> a

timestampFromSeconds :: Integer -> Timestamp

timestampFromUTCTime :: UTCTime -> Timestamp

timestampPlusSeconds :: Timestamp -> Integer -> Timestamp

Add given amount of seconds to a Timestamp.

formatTimestamp :: Timestamp -> Text

Display timestamp in human-readable way as used by Michelson. Uses UTC timezone, though maybe we should take it as an argument.

NB: this will render timestamp with up to seconds precision.

parseTimestamp :: Text -> Maybe Timestamp

Parse textual representation of Timestamp.

timestampQuote :: QuasiQuoter

Quote a value of type Timestamp in yyyy-mm-ddThh:mm:ss[.sss]Z format.

>>> formatTimestamp [timestampQuote| 2019-02-21T16:54:12.2344523Z |]
"2019-02-21T16:54:12Z"

Inspired by 'time-quote' library.

getCurrentTime :: IO Timestamp

Return current time as Timestamp.

farFuture :: Timestamp

Timestamp which is always greater than result of getCurrentTime.

farPast :: Timestamp

Timestamp which is always less than result of getCurrentTime.