Tezos.Crypto

Description

Cryptographic primitives used in Tezos.

Cryptographic primitive types

newtype PublicKey

ED25519 public cryptographic key.

Constructors

PublicKey

Fields

unPublicKey :: PublicKey

data SecretKey

ED25519 secret cryptographic key.

newtype Signature

ED25519 cryptographic signature.

Constructors

Signature

Fields

unSignature :: Signature

newtype KeyHash

b58check of a public key.

Constructors

KeyHash

Fields

unKeyHash :: ByteString

detSecretKey :: ByteString -> SecretKey

Deterministicaly generate a secret key from seed.

toPublic :: SecretKey -> PublicKey

Create a public key from a secret key.

Formatting

data CryptoParseError

Error that can happen during parsing of cryptographic primitive types.

Constructors

CryptoParseWrongBase58Check

CryptoParseWrongTag !ByteString

CryptoParseCryptoError CryptoError

formatPublicKey :: PublicKey -> Text

mformatPublicKey :: PublicKey -> MText

parsePublicKey :: Text -> Either CryptoParseError PublicKey

mkPublicKey :: ByteArrayAccess ba => ba -> Either Text PublicKey

formatSecretKey :: SecretKey -> Text

parseSecretKey :: Text -> Either CryptoParseError SecretKey

formatSignature :: Signature -> Text

mformatSignature :: Signature -> MText

parseSignature :: Text -> Either CryptoParseError Signature

mkSignature :: ByteArrayAccess ba => ba -> Either Text Signature

formatKeyHash :: KeyHash -> Text

mformatKeyHash :: KeyHash -> MText

parseKeyHash :: Text -> Either CryptoParseError KeyHash

Signing

sign :: SecretKey -> ByteString -> Signature

Sign a message using the secret key.

checkSignature :: PublicKey -> Signature -> ByteString -> Bool

Check that a sequence of bytes has been signed with a given key.

Hashing

hashKey :: PublicKey -> KeyHash

Compute the b58check of a public key hash.

blake2b :: ByteString -> ByteString

Compute a cryptographic hash of a bytestring using the Blake2b_256 cryptographic hash function. It's used by the BLAKE2B instruction in Michelson.

blake2b160 :: ByteString -> ByteString

Compute a cryptographic hash of a bytestring using the Blake2b_160 cryptographic hash function.

sha256 :: ByteString -> ByteString

Compute a cryptographic hash of a bytestring using the Sha256 cryptographic hash function.

sha512 :: ByteString -> ByteString

Compute a cryptographic hash of a bytestring using the Sha512 cryptographic hash function.

Utilities

encodeBase58Check :: ByteString -> Text

Encode a bytestring in Base58Check format.

decodeBase58Check :: Text -> Maybe ByteString

Decode a bytestring from Base58Check format.

data B58CheckWithPrefixError

Constructors

B58CheckWithPrefixWrongPrefix ByteString

B58CheckWithPrefixWrongEncoding

decodeBase58CheckWithPrefix :: ByteString -> Text -> Either B58CheckWithPrefixError ByteString

Parse a base58check encoded value expecting some prefix. If the actual prefix matches the expected one, it's stripped of and the resulting payload is returned.